% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{Basic UI skin for TrisomExplorer Applications. Shinydashboard structure with sidebar tabs.}
\usage{
app_ui(
  app_title,
  app_label,
  app_url,
  header_links,
  ui_config,
  app_namespaces,
  input_config
)
}
\arguments{
\item{app_title}{- string - page title for application (see shinydashboardPlus::dashboardPage)}

\item{app_label}{- string - header label to be shown (see shinydashboardPlus::dashboardHeader)}

\item{app_url}{- string - target URL for app_label clicks}

\item{header_links}{- list - list of header tiles / links (see CUSOMShinyHelpers::createApplicationLinks)}

\item{ui_config}{- tibble - configuration values used to dynamically build sidebar tab / content 
see TrisomExploreR::create_tab_items for more detail}

\item{app_namespaces}{- list - list of namespaces / tabs to be dynamically created 
see TrisomExploreR::create_menu_items for more detail}

\item{input_config}{- list - list of input values to be passed to UI modules to hydrate input widgets on init}
}
\value{
shinydashbaordPlus dashboardPage UI element
}
\description{
Basic UI skin for TrisomExplorer Applications. Shinydashboard structure with sidebar tabs.
}
