% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_DatabaseManagers.R
\name{AzureRemoteDataFileManager}
\alias{AzureRemoteDataFileManager}
\title{R6 Class to download remote blob files from Azure Storage}
\description{
download remote blob files from Azure Storage
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{local_data_directory}}{- string - defaults to `data` - path to download remote files locally}

\item{\code{files_downloaded}}{- logical - are the remote files downloaded?}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AzureRemoteDataFileManager-new}{\code{AzureRemoteDataFileManager$new()}}
\item \href{#method-AzureRemoteDataFileManager-download_blob_files}{\code{AzureRemoteDataFileManager$download_blob_files()}}
\item \href{#method-AzureRemoteDataFileManager-get_file_group_directory}{\code{AzureRemoteDataFileManager$get_file_group_directory()}}
\item \href{#method-AzureRemoteDataFileManager-clone}{\code{AzureRemoteDataFileManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzureRemoteDataFileManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-AzureRemoteDataFileManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of AzureRemoteDataFileManager object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzureRemoteDataFileManager$new(
  account_name,
  key,
  container_name,
  local_data_directory = "data",
  refresh = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{account_name}}{- string - Azure Storage Account name}

\item{\code{key}}{- string - authentication key for Azure storage account}

\item{\code{container_name}}{- string - name of target BLOB container}

\item{\code{local_data_directory}}{- string - defaults to `data` - path to donwload remote files locally}

\item{\code{refresh}}{- logical - force a refresh of locally dowloaded data?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `AzureRemoteDataFileManager` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzureRemoteDataFileManager-download_blob_files"></a>}}
\if{latex}{\out{\hypertarget{method-AzureRemoteDataFileManager-download_blob_files}{}}}
\subsection{Method \code{download_blob_files()}}{
download all remote files locally
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzureRemoteDataFileManager$download_blob_files()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzureRemoteDataFileManager-get_file_group_directory"></a>}}
\if{latex}{\out{\hypertarget{method-AzureRemoteDataFileManager-get_file_group_directory}{}}}
\subsection{Method \code{get_file_group_directory()}}{
Helper function to get fully qualified directory name by file_group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzureRemoteDataFileManager$get_file_group_directory(file_group)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file_group}}{- string - name of target file group}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AzureRemoteDataFileManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AzureRemoteDataFileManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AzureRemoteDataFileManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
