% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{bind_events}
\alias{bind_events}
\title{Utility server function to sync input values to appropriate R6 class}
\usage{
bind_events(ids, r6, session, parent_input)
}
\arguments{
\item{ids}{- character vector of input ids to sync to R6 class fields}

\item{r6}{- target R6 class instance}

\item{session}{- shiny session - session for inputs to sync}

\item{parent_input}{- shiny input object from target session}
}
\description{
Utility server function to sync input values to appropriate R6 class
}
