% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_biospecimens_cohort_explorer.R
\name{biospecimens_cohort_explorer_inputs_ui}
\alias{biospecimens_cohort_explorer_inputs_ui}
\title{Create input widgets for TrisomExploreR biospecimen feature analysis}
\usage{
biospecimens_cohort_explorer_inputs_ui(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create input widgets for TrisomExploreR biospecimen feature analysis
}
