% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_cell_type_analysis.R
\name{cell_type_analysis_server}
\alias{cell_type_analysis_server}
\title{Server-side logic / processing for TrisomExploreR cell type analysis}
\usage{
cell_type_analysis_server(id, r6)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic to be utilized by all sub-modules}
}
\description{
Server-side logic / processing for TrisomExploreR cell type analysis
}
