% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_cell_type_analysis.R
\name{cell_type_analysis_ui}
\alias{cell_type_analysis_ui}
\title{Create standard layout / skeleton / sub-modules for TrisomExploreR cell type analysis}
\usage{
cell_type_analysis_ui(id, ...)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{...}{dots - additional arguments (if any) to be passed to sub-modules}
}
\value{
ui module
}
\description{
Create standard layout / skeleton / sub-modules for TrisomExploreR cell type analysis
}
