% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_cell_type_analysis.R
\name{cell_type_inputs_server}
\alias{cell_type_inputs_server}
\title{Server-side logic / processing for TrisomExploreR cell type analysis inputs}
\usage{
cell_type_inputs_server(id, r6)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic for inputs}
}
\description{
Server-side logic / processing for TrisomExploreR cell type analysis inputs
}
