% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_cell_type_analysis.R
\name{cell_type_inputs_ui}
\alias{cell_type_inputs_ui}
\title{Create standard inputs widgets for TrisomExploreR cell type analysis}
\usage{
cell_type_inputs_ui(id, input_config)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{input_config}{list - list of default values for various input widgets}
}
\value{
ui module
}
\description{
Create standard inputs widgets for TrisomExploreR cell type analysis
}
