% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_cell_type.R
\name{cell_type_plot_server}
\alias{cell_type_plot_server}
\title{Server logic for TrisomExploreR celltype boxplots}
\usage{
cell_type_plot_server(id, r6)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic}
}
\description{
Server logic for TrisomExploreR celltype boxplots
}
