% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_cell_type.R
\name{cell_type_plot_ui}
\alias{cell_type_plot_ui}
\title{Create celltype boxplots for TrisomExploreR cell type analysis}
\usage{
cell_type_plot_ui(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create celltype boxplots for TrisomExploreR cell type analysis
}
