% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_CellTypeAnalysis.R
\name{CellTypeAnalysisManager}
\alias{CellTypeAnalysisManager}
\title{R6 Class to manage Cell Type Analysis Modules}
\description{
R6 Class to manage analysis within the Cell Type module.
Retains chosen values from inputs, performs data retrieval, analysis, and plotting.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{applicationName}}{- string - name of application}

\item{\code{namespace}}{- string - namespace for this instance}

\item{\code{remoteDB}}{- R6 class to manage remote database queries}

\item{\code{localDB}}{- R6 class to manage local database queries}

\item{\code{analysisVariable}}{- string - target feature for analysis in this instance-namespace}

\item{\code{analysisVariableLabel}}{- string - label for analysis variable}

\item{\code{analysisType}}{- string - type of analysis - should be one of Categorical or Continuous}

\item{\code{analytesLabel}}{- string - label to be used to refer to set of anlaysis variables}

\item{\code{groupBaselineLabel}}{- string - if analysis type if continous, label for baseline level (Control, Female, Without Condition, etc.)}

\item{\code{FoldChangeVar}}{- string - name of variable indicating fold change or difference (log2FoldChange)}

\item{\code{SignificanceVariable}}{- string - name of variable indiciating significance value (p-value)}

\item{\code{Platform}}{- string vector - Platform values chosen for analysis}

\item{\code{CellType}}{- string vector - Cell Type values chosen for analysis}

\item{\code{Sex}}{- string vector - Sex values chosen for analysis}

\item{\code{Age}}{- numeric vector - Age values chosen fo analysis}

\item{\code{StatTest}}{- string - name of statistical test to apply for analysis (Linear Model, etc.)}

\item{\code{Covariates}}{- string vector - names of features to include as covariates in Linear Model analysis}

\item{\code{AdjustmentMethod}}{- string - name of multiple hypothesis correction method to apply to statistical output}

\item{\code{Analyte}}{- character vector - name(s) of Analytes chosen for analysis}

\item{\code{AnalyteData}}{- tibble - sample level data for chosen analyte(s)}

\item{\code{cids}}{- character vector - control id values chosen to highlight on plot}

\item{\code{tids}}{- character vector - trisomy 21 id values chosen to highlight on plot}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CellTypeAnalysisManager-new}{\code{CellTypeAnalysisManager$new()}}
\item \href{#method-CellTypeAnalysisManager-getCellTypes}{\code{CellTypeAnalysisManager$getCellTypes()}}
\item \href{#method-CellTypeAnalysisManager-getAnalytes}{\code{CellTypeAnalysisManager$getAnalytes()}}
\item \href{#method-CellTypeAnalysisManager-getCovariateChoices}{\code{CellTypeAnalysisManager$getCovariateChoices()}}
\item \href{#method-CellTypeAnalysisManager-getAnalyteData}{\code{CellTypeAnalysisManager$getAnalyteData()}}
\item \href{#method-CellTypeAnalysisManager-getPlot}{\code{CellTypeAnalysisManager$getPlot()}}
\item \href{#method-CellTypeAnalysisManager-clone}{\code{CellTypeAnalysisManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CellTypeAnalysisManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-CellTypeAnalysisManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new CellTypeAnalysisManager object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CellTypeAnalysisManager$new(
  applicationName,
  id,
  namespace_config,
  remoteDB,
  localDB
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{applicationName}}{string - applicationName}

\item{\code{id}}{string - namespace for this class}

\item{\code{namespace_config}}{tibble - configuration values for this namespace instance of the object}

\item{\code{remoteDB}}{R6 class to manage remote database queries}

\item{\code{localDB}}{R6 class to manange local database queries}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `CellTypeAnalysisManager` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CellTypeAnalysisManager-getCellTypes"></a>}}
\if{latex}{\out{\hypertarget{method-CellTypeAnalysisManager-getCellTypes}{}}}
\subsection{Method \code{getCellTypes()}}{
helper function to get cell type list for input
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CellTypeAnalysisManager$getCellTypes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CellTypeAnalysisManager-getAnalytes"></a>}}
\if{latex}{\out{\hypertarget{method-CellTypeAnalysisManager-getAnalytes}{}}}
\subsection{Method \code{getAnalytes()}}{
helper function to get list of genes for input
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CellTypeAnalysisManager$getAnalytes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CellTypeAnalysisManager-getCovariateChoices"></a>}}
\if{latex}{\out{\hypertarget{method-CellTypeAnalysisManager-getCovariateChoices}{}}}
\subsection{Method \code{getCovariateChoices()}}{
helper function to get covariate choices based on namespace
if namespace is Age, remove age, if Sex, remove sex
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CellTypeAnalysisManager$getCovariateChoices()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CellTypeAnalysisManager-getAnalyteData"></a>}}
\if{latex}{\out{\hypertarget{method-CellTypeAnalysisManager-getAnalyteData}{}}}
\subsection{Method \code{getAnalyteData()}}{
Get data for chosen analyte and platform
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CellTypeAnalysisManager$getAnalyteData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CellTypeAnalysisManager-getPlot"></a>}}
\if{latex}{\out{\hypertarget{method-CellTypeAnalysisManager-getPlot}{}}}
\subsection{Method \code{getPlot()}}{
get Cell Type plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CellTypeAnalysisManager$getPlot(.data, ns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- tibble of analyte data to use for plot}

\item{\code{ns}}{- namespace to apply to plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CellTypeAnalysisManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CellTypeAnalysisManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CellTypeAnalysisManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
