% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_clinical.R
\name{clinical_overview_server}
\alias{clinical_overview_server}
\title{overview page for clinical application}
\usage{
clinical_overview_server(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
overview page for clinical application
}
