% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_ClinicalDataAnalysis.R
\name{ClinicalDataAnalysisManager}
\alias{ClinicalDataAnalysisManager}
\title{R6 Class to manage Clincal Data Analysis}
\description{
R6 Class to manage all data analysis within various modules in the Clinical Data application
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{applicationName}}{- string - application name}

\item{\code{namespace}}{- string - namespace for this instance}

\item{\code{remoteDB}}{- R6 Class to manage remote database queries}

\item{\code{localDB}}{- R6 Class to manange local database queries}

\item{\code{ParticipantEncounter}}{- deprecated}

\item{\code{AllParticipants}}{- deprecated}

\item{\code{AllN}}{- numeric - count of all Participants}

\item{\code{Enrollments}}{- tibble of counts of participants by Karyotype}

\item{\code{SampleDetail}}{- tibble of biospecimen detail}

\item{\code{biospecimen_tree}}{- biospecimens in hierarchy / tree stucture}

\item{\code{Biospecimens}}{- selected biospecimen values from UI}

\item{\code{PlatformExperiments}}{- tibble of platform-experiment combinations}

\item{\code{ParticipantPlatformExperiment}}{- tibble of participants, platfoms, and experiments}

\item{\code{AnalysisAvailable}}{- nested list of experiments / studies by platform group}

\item{\code{Karyotypes}}{- character vector - selected karotypes}

\item{\code{Sex}}{- character vector - selected sexes}

\item{\code{Age}}{- numeric vector - selected ages}

\item{\code{OmicsSamples}}{- character vector - selected omics}

\item{\code{Samples}}{- character vector - selected samples}

\item{\code{ParticipantData}}{- tibble of participant level data filtered from inputs}

\item{\code{ConditionClassData}}{- tibble of condition class data}

\item{\code{conditionClasses}}{- character vector - selected condition classes}

\item{\code{conditions}}{- character vector - selected conditions}

\item{\code{conditionClassCounts}}{- tibble - counts of condition classes by Karyotype}

\item{\code{conditionClassSummary}}{- tibble - summary counts of chosen condition classes by Karyotype}

\item{\code{conditionCounts}}{- tibble - counts of conditions by Karyotype}

\item{\code{conditionSummary}}{- tibble - summary counts of conditions by Karyotype}

\item{\code{ConditionClassSexCounts}}{- tibble - counts of condition classes by Sex and Karyotype}

\item{\code{ConditionSexCounts}}{- tibble - counts of conditions by Sex and Karyotype}

\item{\code{selectedAnnotationLevel}}{- string - chosen annotation level for conditions - Condition Class vs Condition}

\item{\code{SelectedConditions}}{- character vector - chosen condition classes or conditions}

\item{\code{upsetPlotData}}{- tibble - wide-form matrix / data frame of condition counts vs. records}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClinicalDataAnalysisManager-new}{\code{ClinicalDataAnalysisManager$new()}}
\item \href{#method-ClinicalDataAnalysisManager-getBiospecimenSampleHierarchy}{\code{ClinicalDataAnalysisManager$getBiospecimenSampleHierarchy()}}
\item \href{#method-ClinicalDataAnalysisManager-getSelectedBiospecimenList}{\code{ClinicalDataAnalysisManager$getSelectedBiospecimenList()}}
\item \href{#method-ClinicalDataAnalysisManager-getParticipantData}{\code{ClinicalDataAnalysisManager$getParticipantData()}}
\item \href{#method-ClinicalDataAnalysisManager-getSexesOverviewPlot}{\code{ClinicalDataAnalysisManager$getSexesOverviewPlot()}}
\item \href{#method-ClinicalDataAnalysisManager-getAgeDistributionOverviewPlot}{\code{ClinicalDataAnalysisManager$getAgeDistributionOverviewPlot()}}
\item \href{#method-ClinicalDataAnalysisManager-getProbandsPlot}{\code{ClinicalDataAnalysisManager$getProbandsPlot()}}
\item \href{#method-ClinicalDataAnalysisManager-getSamplesAvailablePlot}{\code{ClinicalDataAnalysisManager$getSamplesAvailablePlot()}}
\item \href{#method-ClinicalDataAnalysisManager-getOmicsSamplesAvailablePlot}{\code{ClinicalDataAnalysisManager$getOmicsSamplesAvailablePlot()}}
\item \href{#method-ClinicalDataAnalysisManager-getRaceEthnicityPlot}{\code{ClinicalDataAnalysisManager$getRaceEthnicityPlot()}}
\item \href{#method-ClinicalDataAnalysisManager-getParticipantStatesPlot}{\code{ClinicalDataAnalysisManager$getParticipantStatesPlot()}}
\item \href{#method-ClinicalDataAnalysisManager-updateSummaryConditionClassCounts}{\code{ClinicalDataAnalysisManager$updateSummaryConditionClassCounts()}}
\item \href{#method-ClinicalDataAnalysisManager-getConditionClassSummary}{\code{ClinicalDataAnalysisManager$getConditionClassSummary()}}
\item \href{#method-ClinicalDataAnalysisManager-updateConditions}{\code{ClinicalDataAnalysisManager$updateConditions()}}
\item \href{#method-ClinicalDataAnalysisManager-updateSummaryConditionCounts}{\code{ClinicalDataAnalysisManager$updateSummaryConditionCounts()}}
\item \href{#method-ClinicalDataAnalysisManager-updateConditionClassSexCounts}{\code{ClinicalDataAnalysisManager$updateConditionClassSexCounts()}}
\item \href{#method-ClinicalDataAnalysisManager-updateConditionSexCounts}{\code{ClinicalDataAnalysisManager$updateConditionSexCounts()}}
\item \href{#method-ClinicalDataAnalysisManager-getGetConditionStackedSexPlot}{\code{ClinicalDataAnalysisManager$getGetConditionStackedSexPlot()}}
\item \href{#method-ClinicalDataAnalysisManager-update_upset_plot_data}{\code{ClinicalDataAnalysisManager$update_upset_plot_data()}}
\item \href{#method-ClinicalDataAnalysisManager-clear_upset_plot_data}{\code{ClinicalDataAnalysisManager$clear_upset_plot_data()}}
\item \href{#method-ClinicalDataAnalysisManager-clone}{\code{ClinicalDataAnalysisManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of a ClinicalDataAnalysisManager object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$new(
  applicationName,
  id,
  namespace_config,
  remoteDB,
  localDB,
  data
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{applicationName}}{string - applicationName}

\item{\code{id}}{string - namespace for this class}

\item{\code{namespace_config}}{tibble - configuration values for this namespace instance of the object}

\item{\code{remoteDB}}{R6 class to manage remote database queries}

\item{\code{localDB}}{R6 class to manange local database queries}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ClinicalDataAnalysisManager` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-getBiospecimenSampleHierarchy"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-getBiospecimenSampleHierarchy}{}}}
\subsection{Method \code{getBiospecimenSampleHierarchy()}}{
get biospecimen hierarchy for input
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$getBiospecimenSampleHierarchy()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-getSelectedBiospecimenList"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-getSelectedBiospecimenList}{}}}
\subsection{Method \code{getSelectedBiospecimenList()}}{
get aggregated HTML string of chosen biospecimens with line breaks per condition
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$getSelectedBiospecimenList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-getParticipantData"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-getParticipantData}{}}}
\subsection{Method \code{getParticipantData()}}{
Get / set participant level data based on chosen filters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$getParticipantData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-getSexesOverviewPlot"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-getSexesOverviewPlot}{}}}
\subsection{Method \code{getSexesOverviewPlot()}}{
Get participant sex distribution plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$getSexesOverviewPlot(.data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- tibble}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-getAgeDistributionOverviewPlot"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-getAgeDistributionOverviewPlot}{}}}
\subsection{Method \code{getAgeDistributionOverviewPlot()}}{
Get Age distribution plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$getAgeDistributionOverviewPlot(.data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- tibble}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-getProbandsPlot"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-getProbandsPlot}{}}}
\subsection{Method \code{getProbandsPlot()}}{
Get Probands  plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$getProbandsPlot(.data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- tibble}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-getSamplesAvailablePlot"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-getSamplesAvailablePlot}{}}}
\subsection{Method \code{getSamplesAvailablePlot()}}{
Get Samples Availble plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$getSamplesAvailablePlot(.data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- tibble}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-getOmicsSamplesAvailablePlot"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-getOmicsSamplesAvailablePlot}{}}}
\subsection{Method \code{getOmicsSamplesAvailablePlot()}}{
Get OMICS Samples Availble plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$getOmicsSamplesAvailablePlot(.data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- tibble}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-getRaceEthnicityPlot"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-getRaceEthnicityPlot}{}}}
\subsection{Method \code{getRaceEthnicityPlot()}}{
Get Race/Ethnicity plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$getRaceEthnicityPlot(.data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- tibble}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-getParticipantStatesPlot"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-getParticipantStatesPlot}{}}}
\subsection{Method \code{getParticipantStatesPlot()}}{
Get Participants States plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$getParticipantStatesPlot(.data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- tibble}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-updateSummaryConditionClassCounts"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-updateSummaryConditionClassCounts}{}}}
\subsection{Method \code{updateSummaryConditionClassCounts()}}{
set summary condition class counts tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$updateSummaryConditionClassCounts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-getConditionClassSummary"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-getConditionClassSummary}{}}}
\subsection{Method \code{getConditionClassSummary()}}{
get summary condition class counts tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$getConditionClassSummary()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-updateConditions"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-updateConditions}{}}}
\subsection{Method \code{updateConditions()}}{
set conditions counts tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$updateConditions()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-updateSummaryConditionCounts"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-updateSummaryConditionCounts}{}}}
\subsection{Method \code{updateSummaryConditionCounts()}}{
set conditions summary counts tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$updateSummaryConditionCounts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-updateConditionClassSexCounts"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-updateConditionClassSexCounts}{}}}
\subsection{Method \code{updateConditionClassSexCounts()}}{
set conditions class sex counts tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$updateConditionClassSexCounts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-updateConditionSexCounts"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-updateConditionSexCounts}{}}}
\subsection{Method \code{updateConditionSexCounts()}}{
set conditions sex counts tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$updateConditionSexCounts()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-getGetConditionStackedSexPlot"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-getGetConditionStackedSexPlot}{}}}
\subsection{Method \code{getGetConditionStackedSexPlot()}}{
get condition sex plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$getGetConditionStackedSexPlot(
  .data,
  type = "Classes"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{tibble}

\item{\code{type}}{- string - condition annotation level}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-update_upset_plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-update_upset_plot_data}{}}}
\subsection{Method \code{update_upset_plot_data()}}{
set condition upset plot data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$update_upset_plot_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-clear_upset_plot_data"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-clear_upset_plot_data}{}}}
\subsection{Method \code{clear_upset_plot_data()}}{
reset condition upset plot data tibble to NULL states
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$clear_upset_plot_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAnalysisManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAnalysisManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAnalysisManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
