% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_ClinicalDataApp.R
\name{ClinicalDataAppManager}
\alias{ClinicalDataAppManager}
\title{R6 Class to manage Clincal Data Application}
\description{
Sets application-wide configurations, module configurations,
analysis configurations and default input values
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{app_config}}{- list - application-wide configrations (applicationid, title, label, etc.)}

\item{\code{module_config}}{- list - module-namespace configurations}

\item{\code{analysis_config}}{- list - namespace-specific analysis options / configurations}

\item{\code{input_config}}{- list - application-wide default input values}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClinicalDataAppManager-new}{\code{ClinicalDataAppManager$new()}}
\item \href{#method-ClinicalDataAppManager-clone}{\code{ClinicalDataAppManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAppManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAppManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of a ClinicalDataAppManager object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAppManager$new(ApplicationId, remoteDB, localDB)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApplicationId}}{string - ApplicationId}

\item{\code{remoteDB}}{R6 class to manage remote database queries}

\item{\code{localDB}}{R6 class to manange local database queries}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ClinicalDataAppManager` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClinicalDataAppManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClinicalDataAppManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClinicalDataAppManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
