% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_cohort_explorer.R
\name{cohort_explorer_inputs_ui}
\alias{cohort_explorer_inputs_ui}
\title{Create input wigets for TrisomExploreR cohort explorer}
\usage{
cohort_explorer_inputs_ui(id, input_config)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{input_config}{- list - list of default values for various input widgets}
}
\description{
Create input wigets for TrisomExploreR cohort explorer
}
