% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cohort_explorer_kpi.R
\name{cohort_explorer_kpi_server}
\alias{cohort_explorer_kpi_server}
\title{Server for KPI outputs for TrisomExploreR cohort explorer}
\usage{
cohort_explorer_kpi_server(id, r6)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic}
}
\description{
Server for KPI outputs for TrisomExploreR cohort explorer
}
