% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cohort_explorer_kpi.R
\name{cohort_explorer_kpi_ui}
\alias{cohort_explorer_kpi_ui}
\title{Create KPI outputs for TrisomExploreR cohort explorer}
\usage{
cohort_explorer_kpi_ui(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create KPI outputs for TrisomExploreR cohort explorer
}
