% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_cohort_explorer.R
\name{cohort_explorer_server}
\alias{cohort_explorer_server}
\title{Server-side logic / processing for TrisomExploreR cohort explorer}
\usage{
cohort_explorer_server(id, r6, ...)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic to be utilized by all sub-modules}

\item{...}{dots - additional arguments (if any) to be passed to sub-modules}
}
\description{
Server-side logic / processing for TrisomExploreR cohort explorer
}
