% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_cohort_explorer.R
\name{cohort_explorer_ui}
\alias{cohort_explorer_ui}
\title{Create layout / skeleton / objects for TrisomExploreR cohort explorer}
\usage{
cohort_explorer_ui(id, ...)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{...}{dots - additional arguments (if any) to be passed to sub-modules}
}
\value{
ui module
}
\description{
Create layout / skeleton / objects for TrisomExploreR cohort explorer
}
