% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_Comorbidity.R
\name{ComorbidityManager}
\alias{ComorbidityManager}
\title{R6 Class to manage Comorbidity Analysis}
\description{
Subclass of Feature Analysis Manager R6 Class to enable comorbidity-specific analysis.
}
\section{Super class}{
\code{\link[TrisomExploreR:FeatureAnalysisManager]{TrisomExploreR::FeatureAnalysisManager}} -> \code{ComorbidityManager}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ComorbidityManager-new}{\code{ComorbidityManager$new()}}
\item \href{#method-ComorbidityManager-getBaseData}{\code{ComorbidityManager$getBaseData()}}
\item \href{#method-ComorbidityManager-clone}{\code{ComorbidityManager$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="addGSEAInputClass"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-addGSEAInputClass'><code>TrisomExploreR::FeatureAnalysisManager$addGSEAInputClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="addInputSpecialClass"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-addInputSpecialClass'><code>TrisomExploreR::FeatureAnalysisManager$addInputSpecialClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="annotate_volcano_point"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-annotate_volcano_point'><code>TrisomExploreR::FeatureAnalysisManager$annotate_volcano_point()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getAnalyteData"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getAnalyteData'><code>TrisomExploreR::FeatureAnalysisManager$getAnalyteData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getAnalytePlot"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getAnalytePlot'><code>TrisomExploreR::FeatureAnalysisManager$getAnalytePlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getConditionTree"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getConditionTree'><code>TrisomExploreR::FeatureAnalysisManager$getConditionTree()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getCovariateChoices"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getCovariateChoices'><code>TrisomExploreR::FeatureAnalysisManager$getCovariateChoices()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getDisabledInputClass"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getDisabledInputClass'><code>TrisomExploreR::FeatureAnalysisManager$getDisabledInputClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getFormattedAnalyteSummaryData"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getFormattedAnalyteSummaryData'><code>TrisomExploreR::FeatureAnalysisManager$getFormattedAnalyteSummaryData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getFormattedVolcanoSummaryData"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getFormattedVolcanoSummaryData'><code>TrisomExploreR::FeatureAnalysisManager$getFormattedVolcanoSummaryData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getGSEAData"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getGSEAData'><code>TrisomExploreR::FeatureAnalysisManager$getGSEAData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getGSEAEnrichmentPlot"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getGSEAEnrichmentPlot'><code>TrisomExploreR::FeatureAnalysisManager$getGSEAEnrichmentPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getGSEAPathwayData"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getGSEAPathwayData'><code>TrisomExploreR::FeatureAnalysisManager$getGSEAPathwayData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getGSEAPlot"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getGSEAPlot'><code>TrisomExploreR::FeatureAnalysisManager$getGSEAPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getGetDataButtonClass"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getGetDataButtonClass'><code>TrisomExploreR::FeatureAnalysisManager$getGetDataButtonClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getHiddenInputClass"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getHiddenInputClass'><code>TrisomExploreR::FeatureAnalysisManager$getHiddenInputClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getKaryotypeChoices"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getKaryotypeChoices'><code>TrisomExploreR::FeatureAnalysisManager$getKaryotypeChoices()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getSelectedConditionList"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getSelectedConditionList'><code>TrisomExploreR::FeatureAnalysisManager$getSelectedConditionList()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getStudies"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getStudies'><code>TrisomExploreR::FeatureAnalysisManager$getStudies()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getStudyLabel"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getStudyLabel'><code>TrisomExploreR::FeatureAnalysisManager$getStudyLabel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getVolcanoMultiSelectText"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getVolcanoMultiSelectText'><code>TrisomExploreR::FeatureAnalysisManager$getVolcanoMultiSelectText()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getVolcanoPlot"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getVolcanoPlot'><code>TrisomExploreR::FeatureAnalysisManager$getVolcanoPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getVolcanoSummaryData"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getVolcanoSummaryData'><code>TrisomExploreR::FeatureAnalysisManager$getVolcanoSummaryData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="get_analyte_plot_method"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-get_analyte_plot_method'><code>TrisomExploreR::FeatureAnalysisManager$get_analyte_plot_method()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="get_analyte_plot_title"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-get_analyte_plot_title'><code>TrisomExploreR::FeatureAnalysisManager$get_analyte_plot_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="get_analyte_plot_x_axis_label"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-get_analyte_plot_x_axis_label'><code>TrisomExploreR::FeatureAnalysisManager$get_analyte_plot_x_axis_label()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="updateAnalyteAttributes"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-updateAnalyteAttributes'><code>TrisomExploreR::FeatureAnalysisManager$updateAnalyteAttributes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="validate_volcano_plot"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-validate_volcano_plot'><code>TrisomExploreR::FeatureAnalysisManager$validate_volcano_plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComorbidityManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-ComorbidityManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of ComorbidityManager object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComorbidityManager$new(
  applicationName,
  id,
  namespace_config,
  remoteDB,
  localDB
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{applicationName}}{string - applicationName}

\item{\code{id}}{string - namespace for this class}

\item{\code{namespace_config}}{tibble - configuration values for this namespace instance of the object}

\item{\code{remoteDB}}{R6 class to manage remote database queries}

\item{\code{localDB}}{R6 class to manange local database queries}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ComorbidityManager` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComorbidityManager-getBaseData"></a>}}
\if{latex}{\out{\hypertarget{method-ComorbidityManager-getBaseData}{}}}
\subsection{Method \code{getBaseData()}}{
set BaseData for comorbidity analysis - custom logic that overrides `getBaseData` method from
 parent `FeatureAnalysisManager` class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComorbidityManager$getBaseData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ComorbidityManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ComorbidityManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ComorbidityManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
