% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_condition_correlates_boxplot.R
\name{condition_correlates_boxplot_server}
\alias{condition_correlates_boxplot_server}
\title{Server for KPI outputs for TrisomExploreR condition correlates analysis
Entire logic is server-side contained within a modal 
only invoked from `gargoyle::watch("show_analyte_plot")`}
\usage{
condition_correlates_boxplot_server(id, r6)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic}
}
\description{
Server for KPI outputs for TrisomExploreR condition correlates analysis
Entire logic is server-side contained within a modal 
only invoked from `gargoyle::watch("show_analyte_plot")`
}
