% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_condition_correlates_boxplot.R
\name{condition_correlates_boxplot_ui}
\alias{condition_correlates_boxplot_ui}
\title{Create boxplot UI for TrisomExploreR condition correlates analysis}
\usage{
condition_correlates_boxplot_ui(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create boxplot UI for TrisomExploreR condition correlates analysis
}
