% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_condition_correlates_heatmap.R
\name{condition_correlates_heatmap_plot_server}
\alias{condition_correlates_heatmap_plot_server}
\title{Server logic for TrisomExploreR condition correlates heatmap plot}
\usage{
condition_correlates_heatmap_plot_server(id, r6)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic}
}
\description{
Server logic for TrisomExploreR condition correlates heatmap plot
}
