% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_condition_correlates_heatmap.R
\name{condition_correlates_heatmap_plot_ui}
\alias{condition_correlates_heatmap_plot_ui}
\title{Create heatmap plot for TrisomExploreR condition correlates analysis}
\usage{
condition_correlates_heatmap_plot_ui(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create heatmap plot for TrisomExploreR condition correlates analysis
}
