% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_condition_correlates.R
\name{condition_correlates_inputs_server}
\alias{condition_correlates_inputs_server}
\title{Server-side logic / processing for TrisomExploreR condition correlates explorer inputs}
\usage{
condition_correlates_inputs_server(id, r6, input_config)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic to be utilized by all sub-modules}

\item{input_config}{- list - list of default values for various input widgets to be used server-side}
}
\description{
Server-side logic / processing for TrisomExploreR condition correlates explorer inputs
}
