% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_condition_correlates.R
\name{condition_correlates_ui}
\alias{condition_correlates_ui}
\title{Create layout / skeleton / objects / submodules for TrisomExploreR condition correlates explorer}
\usage{
condition_correlates_ui(id, ...)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{...}{dots - additional arguments (if any) to be passed to sub-modules}
}
\description{
Create layout / skeleton / objects / submodules for TrisomExploreR condition correlates explorer
}
