% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_condition_frequency_class_sex.R
\name{condition_frequency_class_sex_plot_ui}
\alias{condition_frequency_class_sex_plot_ui}
\title{Create condition class sex plot for TrisomExploreR condition frequency analysis}
\usage{
condition_frequency_class_sex_plot_ui(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create condition class sex plot for TrisomExploreR condition frequency analysis
}
