% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_condition_frequency_class.R
\name{condition_frequency_class_table_server}
\alias{condition_frequency_class_table_server}
\title{Server logic for condition class frequency table output for TrisomExploreR Clinical data analysis}
\usage{
condition_frequency_class_table_server(id, r6)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic}
}
\description{
Server logic for condition class frequency table output for TrisomExploreR Clinical data analysis
}
