% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_condition_frequency_class.R
\name{condition_frequency_class_table_ui}
\alias{condition_frequency_class_table_ui}
\title{Create condition class frequency table output for TrisomExploreR Clinical data analysis}
\usage{
condition_frequency_class_table_ui(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create condition class frequency table output for TrisomExploreR Clinical data analysis
}
