% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_condition_frequency_condition.R
\name{condition_frequency_condition_table_ui}
\alias{condition_frequency_condition_table_ui}
\title{Create condition frequency table output for TrisomExploreR Clinical data analysis}
\usage{
condition_frequency_condition_table_ui(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create condition frequency table output for TrisomExploreR Clinical data analysis
}
