% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_condition_frequency.R
\name{condition_frequency_ui}
\alias{condition_frequency_ui}
\title{Create layout / skeleton / objects for TrisomExploreR condition frequency analysis}
\usage{
condition_frequency_ui(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create layout / skeleton / objects for TrisomExploreR condition frequency analysis
}
