% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_condition_interactions.R
\name{condition_interactions_plot_server}
\alias{condition_interactions_plot_server}
\title{Server logic for TrisomExploreR condition interactions upset plot}
\usage{
condition_interactions_plot_server(id, r6)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic}
}
\description{
Server logic for TrisomExploreR condition interactions upset plot
}
