% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_condition_interactions.R
\name{condition_interactions_plot_ui}
\alias{condition_interactions_plot_ui}
\title{Create upset plot for TrisomExploreR condition interactions analysis}
\usage{
condition_interactions_plot_ui(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create upset plot for TrisomExploreR condition interactions analysis
}
