% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_condition_interactions.R
\name{condition_interactions_ui}
\alias{condition_interactions_ui}
\title{Create layout / skeleton / objects for TrisomExploreR condition interactions}
\usage{
condition_interactions_ui(id, ...)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{...}{dots - additional arguments (if any) to be passed to sub-modules}
}
\description{
Create layout / skeleton / objects for TrisomExploreR condition interactions
}
