% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_ConditionCorrelates.R
\name{ConditionCorrelatesManager}
\alias{ConditionCorrelatesManager}
\title{R6 Class to manage Condition Correlates Analysis}
\description{
Enables analysis of co-occuring conditions by chosen feature
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{applicationName}}{- string - name of application}

\item{\code{namespace}}{- string - namespace for this instance}

\item{\code{remoteDB}}{- R6 class to manage remote database queries}

\item{\code{localDB}}{- R6 class to manage local database queries}

\item{\code{analysisVariable}}{- string - target feature for analysis in this instance-namespace}

\item{\code{analysisVariableLabel}}{- string - label for analysis variable}

\item{\code{FoldChangeVar}}{- string - name of variable indicating fold change or difference (log2FoldChange)}

\item{\code{SignificanceVariable}}{- string - name of variable indiciating significance value (p-value)}

\item{\code{Study}}{- string - selected study}

\item{\code{Platform}}{- string vector - Platform values chosen for analysis}

\item{\code{Experiment}}{- string - chosen experimentID}

\item{\code{QueryAnalyte}}{- string - chosen Query Analyte}

\item{\code{Analyte}}{- string - chosen Comparison Analyte}

\item{\code{MinComorbitityMembership}}{- numeric - minimum number of records with matching condition(s) to include in analysis}

\item{\code{Sex}}{- string vector - Sex values chosen for analysis}

\item{\code{Age}}{- numeric vector - Age values chosen fo analysis}

\item{\code{StatTest}}{- string - name of statistical test to apply for analysis (Linear Model, etc.)}

\item{\code{Covariates}}{- string vector - names of features to include as covariates in Linear Model analysis}

\item{\code{AdjustmentMethod}}{- string - name of multiple hypothesis correction method to apply to statistical output}

\item{\code{AnalytePlotStatAnnotation}}{- string - stat annotation to show above analyte plot}

\item{\code{AnalyteData}}{- tibble - sample level data for chosen analyte(s)}

\item{\code{ParticipantComorbidities}}{- tibble - Participant Comorbidity data}

\item{\code{ComorbiditySummary}}{- tibble -}

\item{\code{Comorbidities}}{- character vector - chosen comborbidities}

\item{\code{VolcanoSummaryData}}{- tibble - Fold Change summary data used for volcano plot}

\item{\code{VolcanoSummaryDataXAxisLabel}}{- string - volcano plot x-axis}

\item{\code{VolcanoSummaryDataYAxisLabel}}{- string - volcano plot y-axis}

\item{\code{VolcanoSummaryMaxFoldChange}}{- numeric - maxiumum abs. value of fold change}

\item{\code{VolcanoPlotTitle}}{- string - title to show above volcano plot}

\item{\code{volcanoTopAnnotationLabel}}{- string - text to show above top annotations on volcano plot}

\item{\code{volcanoPlotExpectedTraceCount}}{- numeric - number of base traces present in the active volcano plot (usually between 1 - 3)}

\item{\code{VolcanoSummaryDataFoldChangeFilter}}{- depreciated?}

\item{\code{volcanoMultiSelectText}}{- string - text to show below volcano plot when multiple analytes are chosen}

\item{\code{HeatmapData}}{- tibble - data to use for heatmap plot when multiple analytes are chosen}

\item{\code{BoxplotData}}{- tibble - data to use for boxplot when a single anlayte is chosen}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConditionCorrelatesManager-new}{\code{ConditionCorrelatesManager$new()}}
\item \href{#method-ConditionCorrelatesManager-addGSEAInputClass}{\code{ConditionCorrelatesManager$addGSEAInputClass()}}
\item \href{#method-ConditionCorrelatesManager-getQueryAnalytes}{\code{ConditionCorrelatesManager$getQueryAnalytes()}}
\item \href{#method-ConditionCorrelatesManager-getVolcanoSummaryData}{\code{ConditionCorrelatesManager$getVolcanoSummaryData()}}
\item \href{#method-ConditionCorrelatesManager-getFormattedVolcanoSummaryData}{\code{ConditionCorrelatesManager$getFormattedVolcanoSummaryData()}}
\item \href{#method-ConditionCorrelatesManager-getParticipantComorbiditiesByExperimentAnalyte}{\code{ConditionCorrelatesManager$getParticipantComorbiditiesByExperimentAnalyte()}}
\item \href{#method-ConditionCorrelatesManager-getComorbidtySummary}{\code{ConditionCorrelatesManager$getComorbidtySummary()}}
\item \href{#method-ConditionCorrelatesManager-getVolcanoPlot}{\code{ConditionCorrelatesManager$getVolcanoPlot()}}
\item \href{#method-ConditionCorrelatesManager-getHeatmapPlot}{\code{ConditionCorrelatesManager$getHeatmapPlot()}}
\item \href{#method-ConditionCorrelatesManager-getFormattedAnalyteSummaryData}{\code{ConditionCorrelatesManager$getFormattedAnalyteSummaryData()}}
\item \href{#method-ConditionCorrelatesManager-getBoxPlotData}{\code{ConditionCorrelatesManager$getBoxPlotData()}}
\item \href{#method-ConditionCorrelatesManager-getBoxPlot}{\code{ConditionCorrelatesManager$getBoxPlot()}}
\item \href{#method-ConditionCorrelatesManager-getFormattedBoxplotData}{\code{ConditionCorrelatesManager$getFormattedBoxplotData()}}
\item \href{#method-ConditionCorrelatesManager-clone}{\code{ConditionCorrelatesManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionCorrelatesManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionCorrelatesManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of ConditionCorrelatesManager object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionCorrelatesManager$new(
  applicationName,
  id,
  namespace_config,
  remoteDB,
  localDB
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{applicationName}}{string - applicationName}

\item{\code{id}}{string - namespace for this class}

\item{\code{namespace_config}}{tibble - configuration values for this namespace instance of the object}

\item{\code{remoteDB}}{R6 class to manage remote database queries}

\item{\code{localDB}}{R6 class to manange local database queries}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ConditionCorrelatesManager` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionCorrelatesManager-addGSEAInputClass"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionCorrelatesManager-addGSEAInputClass}{}}}
\subsection{Method \code{addGSEAInputClass()}}{
return hidden /shown / disabled / enabled class for GSEA button based on
 chosen comparison platform and whether or not Volcano Plot is rendered
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionCorrelatesManager$addGSEAInputClass()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionCorrelatesManager-getQueryAnalytes"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionCorrelatesManager-getQueryAnalytes}{}}}
\subsection{Method \code{getQueryAnalytes()}}{
Set Volcano Summary Data along with other volcano plot properties


Retrieve list of query analytes for a given platform and experiment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionCorrelatesManager$getQueryAnalytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionCorrelatesManager-getVolcanoSummaryData"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionCorrelatesManager-getVolcanoSummaryData}{}}}
\subsection{Method \code{getVolcanoSummaryData()}}{
Set Volcano Summary Data
 Also sets ParticipantComorbidities, ComorbiditySummary, Comborbidities, AnalyteData, and volcano plot properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionCorrelatesManager$getVolcanoSummaryData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionCorrelatesManager-getFormattedVolcanoSummaryData"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionCorrelatesManager-getFormattedVolcanoSummaryData}{}}}
\subsection{Method \code{getFormattedVolcanoSummaryData()}}{
Get formatted Volcano Summary Data for data table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionCorrelatesManager$getFormattedVolcanoSummaryData(.data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- tibble - data to be formatted}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionCorrelatesManager-getParticipantComorbiditiesByExperimentAnalyte"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionCorrelatesManager-getParticipantComorbiditiesByExperimentAnalyte}{}}}
\subsection{Method \code{getParticipantComorbiditiesByExperimentAnalyte()}}{
Get Participant Comorbidities for a given experiment and analyte
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionCorrelatesManager$getParticipantComorbiditiesByExperimentAnalyte(
  Experiment,
  QueryAnalyte
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{Experiment}}{- string}

\item{\code{QueryAnalyte}}{- string}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionCorrelatesManager-getComorbidtySummary"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionCorrelatesManager-getComorbidtySummary}{}}}
\subsection{Method \code{getComorbidtySummary()}}{
Get Comoborbidity Summary data - appends each comorbidity label with {positive} {negative} counts
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionCorrelatesManager$getComorbidtySummary(
  .data,
  MinComorbitityMembership
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- tibble - tibble of participant comborbidity data}

\item{\code{MinComorbitityMembership}}{- numeric - minimum threshold of diagnosed to include in analysis}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionCorrelatesManager-getVolcanoPlot"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionCorrelatesManager-getVolcanoPlot}{}}}
\subsection{Method \code{getVolcanoPlot()}}{
Get Volcano Plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionCorrelatesManager$getVolcanoPlot(.data, ns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- tibble - data for plot}

\item{\code{ns}}{- namespace for plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionCorrelatesManager-getHeatmapPlot"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionCorrelatesManager-getHeatmapPlot}{}}}
\subsection{Method \code{getHeatmapPlot()}}{
Get Heatmap Plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionCorrelatesManager$getHeatmapPlot(.data, ns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- tibble - data for plot}

\item{\code{ns}}{- namespace for plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionCorrelatesManager-getFormattedAnalyteSummaryData"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionCorrelatesManager-getFormattedAnalyteSummaryData}{}}}
\subsection{Method \code{getFormattedAnalyteSummaryData()}}{
Get Formatted Analyte Summary data for data table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionCorrelatesManager$getFormattedAnalyteSummaryData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionCorrelatesManager-getBoxPlotData"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionCorrelatesManager-getBoxPlotData}{}}}
\subsection{Method \code{getBoxPlotData()}}{
set data for boxplot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionCorrelatesManager$getBoxPlotData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionCorrelatesManager-getBoxPlot"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionCorrelatesManager-getBoxPlot}{}}}
\subsection{Method \code{getBoxPlot()}}{
get boxplot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionCorrelatesManager$getBoxPlot(.data, ns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- tibble - data used for boxplot}

\item{\code{ns}}{- namespace to apply to plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionCorrelatesManager-getFormattedBoxplotData"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionCorrelatesManager-getFormattedBoxplotData}{}}}
\subsection{Method \code{getFormattedBoxplotData()}}{
get formatted data for boxplot data table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionCorrelatesManager$getFormattedBoxplotData(.data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- tibble - data to format}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConditionCorrelatesManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConditionCorrelatesManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConditionCorrelatesManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
