% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_conditions_feature_analysis.R
\name{conditions_feature_analysis_inputs_server}
\alias{conditions_feature_analysis_inputs_server}
\title{Server-side logic / processing for TrisomExploreR conditions feature analysis inputs}
\usage{
conditions_feature_analysis_inputs_server(id, r6, input_config, parent)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic for inputs}

\item{input_config}{- list - list of default values for various input widgets to be used server-side}

\item{parent}{- session object - parent session}
}
\description{
Server-side logic / processing for TrisomExploreR conditions feature analysis inputs
}
