% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_conditions_feature_analysis.R
\name{conditions_feature_analysis_inputs_ui}
\alias{conditions_feature_analysis_inputs_ui}
\title{Create input widgets for TrisomExploreR conditions feature analysis / additional inputs for co-occuring condition analysis}
\usage{
conditions_feature_analysis_inputs_ui(id, input_config)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{input_config}{- list - list of default values for various input widgets}
}
\description{
Create input widgets for TrisomExploreR conditions feature analysis / additional inputs for co-occuring condition analysis
}
