% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_correlates.R
\name{correlates_inputs_ui}
\alias{correlates_inputs_ui}
\title{Create input widgets for TrisomExploreR cross-omics correlates analysis}
\usage{
correlates_inputs_ui(id, input_config)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{input_config}{- list - list of default values for various input widgets}
}
\description{
Create input widgets for TrisomExploreR cross-omics correlates analysis
}
