% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_correlates.R
\name{correlates_ui}
\alias{correlates_ui}
\title{Create layout / skeleton / sub-modules for TrisomExploreR cross-omics correlates analysis}
\usage{
correlates_ui(id, ...)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{...}{dots - additional arguments (if any) to be passed to sub-modules}
}
\description{
Create layout / skeleton / sub-modules for TrisomExploreR cross-omics correlates analysis
}
