% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{create_menu_item}
\alias{create_menu_item}
\title{Utility UI function to create `shinydashboard::menuItem` from config}
\usage{
create_menu_item(id, ui_config)
}
\arguments{
\item{id}{- string - namespace for this instance of `menuItem` object}

\item{ui_config}{- list of ui configurations to build `menuItem` object}
}
\description{
Utility UI function to create `shinydashboard::menuItem` from config
}
