% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{create_menu_items}
\alias{create_menu_items}
\title{Utility UI function to dynamically create `shinydashboard::sidebarMenu` items from config}
\usage{
create_menu_items(namespaces, ui_config)
}
\arguments{
\item{namespaces}{- list of namespaces / tab items}

\item{ui_config}{- list of ui configurations for menu item (`text`,`name`,`icon`)}
}
\description{
Utility UI function to dynamically create `shinydashboard::sidebarMenu` items from config
}
