% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{create_tab_item}
\alias{create_tab_item}
\title{Utility UI function to create `shinydashboard::tabItem` from config}
\usage{
create_tab_item(id, ui_config, input_config)
}
\arguments{
\item{id}{- string - namespace for this instance of `tabItem`}

\item{ui_config}{- list of ui configuration for this instance of `tabItem`}

\item{input_config}{- list of input configurations to be passed to appropriate ui module}
}
\description{
Utility UI function to create `shinydashboard::tabItem` from config
}
