% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_utils.R
\name{create_tab_items}
\alias{create_tab_items}
\title{Utility UI function to dynamically create `shinydashboard::tabItems` from config}
\usage{
create_tab_items(namespaces, ui_config, input_config)
}
\arguments{
\item{namespaces}{- list of namespaces / tab items}

\item{ui_config}{- list of ui configurations to build `tabItem`}

\item{input_config}{- list of ui input configurations to be passed as argument to ui module}
}
\description{
Utility UI function to dynamically create `shinydashboard::tabItems` from config
}
