% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_feature_analysis_analyte.R
\name{feature_analysis_analyte_plot_server}
\alias{feature_analysis_analyte_plot_server}
\title{Server logic / processing for TrisomExploreR feature analysis analyte plot}
\usage{
feature_analysis_analyte_plot_server(id, r6)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic}
}
\description{
Server logic / processing for TrisomExploreR feature analysis analyte plot
}
