% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_feature_analysis_analyte.R
\name{feature_analysis_analyte_plot_ui}
\alias{feature_analysis_analyte_plot_ui}
\title{Create analyte plot for TrisomExploreR feature analysis}
\usage{
feature_analysis_analyte_plot_ui(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create analyte plot for TrisomExploreR feature analysis
}
