% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_feature_analysis_analyte.R
\name{feature_analysis_analyte_summary_data_server}
\alias{feature_analysis_analyte_summary_data_server}
\title{Server logic for Analyte table output for TrisomExploreR feature analysis}
\usage{
feature_analysis_analyte_summary_data_server(id, r6)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic}
}
\description{
Server logic for Analyte table output for TrisomExploreR feature analysis
}
