% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_feature_analysis_analyte.R
\name{feature_analysis_analyte_summary_data_ui}
\alias{feature_analysis_analyte_summary_data_ui}
\title{Create Analyte table output for TrisomExploreR feature analysis}
\usage{
feature_analysis_analyte_summary_data_ui(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create Analyte table output for TrisomExploreR feature analysis
}
