% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_GSEA_analysis.R
\name{feature_analysis_GSEA_plot_server}
\alias{feature_analysis_GSEA_plot_server}
\title{Server logic / processing for TrisomExploreR GSEA plot}
\usage{
feature_analysis_GSEA_plot_server(id, r6, parent)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic}

\item{parent}{shiny session - parent shiny session}
}
\description{
Server logic / processing for TrisomExploreR GSEA plot
}
