% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_GSEA_analysis.R
\name{feature_analysis_GSEA_plot_ui}
\alias{feature_analysis_GSEA_plot_ui}
\title{Create GSEA plot for TrisomExploreR GSEA pathway analysis}
\usage{
feature_analysis_GSEA_plot_ui(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create GSEA plot for TrisomExploreR GSEA pathway analysis
}
