% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_GSEA_analysis.R
\name{feature_analysis_GSEA_summary_data_server}
\alias{feature_analysis_GSEA_summary_data_server}
\title{Server logic for GSEA pathway table output for TrisomExploreR pathway analysis}
\usage{
feature_analysis_GSEA_summary_data_server(id, r6, parent)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic}

\item{parent}{- shiny session object - parent session}
}
\description{
Server logic for GSEA pathway table output for TrisomExploreR pathway analysis
}
