% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_GSEA_analysis.R
\name{feature_analysis_GSEA_summary_data_ui}
\alias{feature_analysis_GSEA_summary_data_ui}
\title{Create GSEA pathway table output for TrisomExploreR pathway analysis}
\usage{
feature_analysis_GSEA_summary_data_ui(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create GSEA pathway table output for TrisomExploreR pathway analysis
}
