% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureAnalysisManager.R
\name{FeatureAnalysisManager}
\alias{FeatureAnalysisManager}
\title{R6 Class to}
\description{
R6 Class to

R6 Class to
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{applicationName}}{- string - application name}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FeatureAnalysisManager-new}{\code{FeatureAnalysisManager$new()}}
\item \href{#method-FeatureAnalysisManager-getGetDataButtonClass}{\code{FeatureAnalysisManager$getGetDataButtonClass()}}
\item \href{#method-FeatureAnalysisManager-getKaryotypeChoices}{\code{FeatureAnalysisManager$getKaryotypeChoices()}}
\item \href{#method-FeatureAnalysisManager-getConditionTree}{\code{FeatureAnalysisManager$getConditionTree()}}
\item \href{#method-FeatureAnalysisManager-getSelectedConditionList}{\code{FeatureAnalysisManager$getSelectedConditionList()}}
\item \href{#method-FeatureAnalysisManager-getCovariateChoices}{\code{FeatureAnalysisManager$getCovariateChoices()}}
\item \href{#method-FeatureAnalysisManager-getDisabledInputClass}{\code{FeatureAnalysisManager$getDisabledInputClass()}}
\item \href{#method-FeatureAnalysisManager-getHiddenInputClass}{\code{FeatureAnalysisManager$getHiddenInputClass()}}
\item \href{#method-FeatureAnalysisManager-addInputSpecialClass}{\code{FeatureAnalysisManager$addInputSpecialClass()}}
\item \href{#method-FeatureAnalysisManager-addGSEAInputClass}{\code{FeatureAnalysisManager$addGSEAInputClass()}}
\item \href{#method-FeatureAnalysisManager-showGSEAButton}{\code{FeatureAnalysisManager$showGSEAButton()}}
\item \href{#method-FeatureAnalysisManager-enableGSEAButton}{\code{FeatureAnalysisManager$enableGSEAButton()}}
\item \href{#method-FeatureAnalysisManager-getBaseData}{\code{FeatureAnalysisManager$getBaseData()}}
\item \href{#method-FeatureAnalysisManager-getVolcanoSummaryData}{\code{FeatureAnalysisManager$getVolcanoSummaryData()}}
\item \href{#method-FeatureAnalysisManager-getFormattedVolcanoSummaryData}{\code{FeatureAnalysisManager$getFormattedVolcanoSummaryData()}}
\item \href{#method-FeatureAnalysisManager-getVolcanoPlot}{\code{FeatureAnalysisManager$getVolcanoPlot()}}
\item \href{#method-FeatureAnalysisManager-updateAnalyteAttributes}{\code{FeatureAnalysisManager$updateAnalyteAttributes()}}
\item \href{#method-FeatureAnalysisManager-getAnalyteData}{\code{FeatureAnalysisManager$getAnalyteData()}}
\item \href{#method-FeatureAnalysisManager-getAnalytePlot}{\code{FeatureAnalysisManager$getAnalytePlot()}}
\item \href{#method-FeatureAnalysisManager-getFormattedAnalyteSummaryData}{\code{FeatureAnalysisManager$getFormattedAnalyteSummaryData()}}
\item \href{#method-FeatureAnalysisManager-getGSEAData}{\code{FeatureAnalysisManager$getGSEAData()}}
\item \href{#method-FeatureAnalysisManager-getGSEAPlot}{\code{FeatureAnalysisManager$getGSEAPlot()}}
\item \href{#method-FeatureAnalysisManager-getGSEAPathwayData}{\code{FeatureAnalysisManager$getGSEAPathwayData()}}
\item \href{#method-FeatureAnalysisManager-getGSEAEnrichmentPlot}{\code{FeatureAnalysisManager$getGSEAEnrichmentPlot()}}
\item \href{#method-FeatureAnalysisManager-clone}{\code{FeatureAnalysisManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$new(
  applicationName,
  id,
  namespace_config,
  remoteDB,
  localDB
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getGetDataButtonClass"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getGetDataButtonClass}{}}}
\subsection{Method \code{getGetDataButtonClass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getGetDataButtonClass()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getKaryotypeChoices"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getKaryotypeChoices}{}}}
\subsection{Method \code{getKaryotypeChoices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getKaryotypeChoices(karyotypes)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getConditionTree"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getConditionTree}{}}}
\subsection{Method \code{getConditionTree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getConditionTree(conditions)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getSelectedConditionList"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getSelectedConditionList}{}}}
\subsection{Method \code{getSelectedConditionList()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getSelectedConditionList()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getCovariateChoices"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getCovariateChoices}{}}}
\subsection{Method \code{getCovariateChoices()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getCovariateChoices()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getDisabledInputClass"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getDisabledInputClass}{}}}
\subsection{Method \code{getDisabledInputClass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getDisabledInputClass(inputName)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getHiddenInputClass"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getHiddenInputClass}{}}}
\subsection{Method \code{getHiddenInputClass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getHiddenInputClass(inputName)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-addInputSpecialClass"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-addInputSpecialClass}{}}}
\subsection{Method \code{addInputSpecialClass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$addInputSpecialClass(
  inputName,
  class = c("disabled", "hide")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-addGSEAInputClass"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-addGSEAInputClass}{}}}
\subsection{Method \code{addGSEAInputClass()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$addGSEAInputClass()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-showGSEAButton"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-showGSEAButton}{}}}
\subsection{Method \code{showGSEAButton()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$showGSEAButton(button_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-enableGSEAButton"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-enableGSEAButton}{}}}
\subsection{Method \code{enableGSEAButton()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$enableGSEAButton(button_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getBaseData"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getBaseData}{}}}
\subsection{Method \code{getBaseData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getBaseData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getVolcanoSummaryData"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getVolcanoSummaryData}{}}}
\subsection{Method \code{getVolcanoSummaryData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getVolcanoSummaryData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getFormattedVolcanoSummaryData"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getFormattedVolcanoSummaryData}{}}}
\subsection{Method \code{getFormattedVolcanoSummaryData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getFormattedVolcanoSummaryData(.data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getVolcanoPlot"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getVolcanoPlot}{}}}
\subsection{Method \code{getVolcanoPlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getVolcanoPlot(.data, ns)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-updateAnalyteAttributes"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-updateAnalyteAttributes}{}}}
\subsection{Method \code{updateAnalyteAttributes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$updateAnalyteAttributes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getAnalyteData"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getAnalyteData}{}}}
\subsection{Method \code{getAnalyteData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getAnalyteData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getAnalytePlot"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getAnalytePlot}{}}}
\subsection{Method \code{getAnalytePlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getAnalytePlot(.data, ns)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getFormattedAnalyteSummaryData"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getFormattedAnalyteSummaryData}{}}}
\subsection{Method \code{getFormattedAnalyteSummaryData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getFormattedAnalyteSummaryData(.data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getGSEAData"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getGSEAData}{}}}
\subsection{Method \code{getGSEAData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getGSEAData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getGSEAPlot"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getGSEAPlot}{}}}
\subsection{Method \code{getGSEAPlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getGSEAPlot(.data, ns)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getGSEAPathwayData"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getGSEAPathwayData}{}}}
\subsection{Method \code{getGSEAPathwayData()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getGSEAPathwayData(pathName)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-getGSEAEnrichmentPlot"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-getGSEAEnrichmentPlot}{}}}
\subsection{Method \code{getGSEAEnrichmentPlot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$getGSEAEnrichmentPlot(.data, ns)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FeatureAnalysisManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FeatureAnalysisManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FeatureAnalysisManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
