% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{get_object_name_from_namespace_session}
\alias{get_object_name_from_namespace_session}
\title{Helper function to find and return fully-qualified / namespaced object name}
\usage{
get_object_name_from_namespace_session(session, namespace, object_name)
}
\arguments{
\item{session}{- shiny session object}

\item{namespace}{- string - namespace for target object}

\item{object_name}{- string - name of target object}
}
\description{
Helper function to find and return fully-qualified / namespaced object name
}
