% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureAnalysisManager.R
\name{getAnalytePlotTitle}
\alias{getAnalytePlotTitle}
\title{small helper function to generate plot title based on criteria}
\usage{
getAnalytePlotTitle(analysisVariable, plotMethod, analyte, groupVariable)
}
\arguments{
\item{analysisVariable}{- string - name of analysis variable}

\item{plotMethod}{- string - one of boxplot, scatterplot, or heatmap}

\item{analyte}{- string vector - vector of chosen analytes}

\item{groupVariable}{- string - grouping variable used in analysis (Karyotype)}
}
\value{
string
}
\description{
small helper function to generate plot title based on criteria
}
