% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_FeatureAnalysis.R
\name{getAnalytePlotXAxisLabel}
\alias{getAnalytePlotXAxisLabel}
\title{small helper function to generate x-axis label for analyte plot}
\usage{
getAnalytePlotXAxisLabel(namespace, analysis_variable)
}
\arguments{
\item{namespace}{- string - namespace}

\item{analysis_variable}{- string - name of analysis variable}
}
\description{
small helper function to generate x-axis label for analyte plot
}
