% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_GSEA_analysis.R
\name{GSEA_analysis_inputs_server}
\alias{GSEA_analysis_inputs_server}
\title{Server-side logic / processing for TrisomExploreR GSEA pathway analysis inputs}
\usage{
GSEA_analysis_inputs_server(id, r6, parent)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic for inputs}

\item{parent}{- session object - parent session}
}
\description{
Server-side logic / processing for TrisomExploreR GSEA pathway analysis inputs
}
