% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{GSEA_hallmarks}
\alias{GSEA_hallmarks}
\title{GSEA_hallmarks}
\format{
A list
}
\usage{
GSEA_hallmarks
}
\description{
A list of 50 Hallmark gene sets / pathways for Gene Set Enrichment Analysis. Data was sourced from the UC San Diego Gene Set Enrichment Analysis website:
\url{https://www.gsea-msigdb.org/gsea/msigdb/human/collections.jsp#H}
}
\details{
\itemize{
 \item GSEA_hallmarks - list of 50 named gene set / pathway along with a collection of genes listed in a delimited string
}
}
\keyword{datasets}
