% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_immunemap_correlates.R
\name{immunemap_correlates_ui}
\alias{immunemap_correlates_ui}
\title{Create layout / skeleton / sub-modules for TrisomExploreR Immune Maps specific cross-omics correlates analysis
Follows standard layout for correlates analysis, except Immune Map inputs are different}
\usage{
immunemap_correlates_ui(id, ...)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{...}{dots - additional arguments (if any) to be passed to sub-modules}
}
\description{
Create layout / skeleton / sub-modules for TrisomExploreR Immune Maps specific cross-omics correlates analysis
Follows standard layout for correlates analysis, except Immune Map inputs are different
}
