% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_immunemap_feature_analysis.R
\name{immunemap_feature_analysis_ui}
\alias{immunemap_feature_analysis_ui}
\title{Create standard layout / skeleton / sub-modules for TrisomExploreR Immune Map specific feature analysis}
\usage{
immunemap_feature_analysis_ui(id, ...)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{...}{dots - additional arguments (if any) to be passed to sub-modules}
}
\description{
Create standard layout / skeleton / sub-modules for TrisomExploreR Immune Map specific feature analysis
}
