% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_immunemap.R
\name{immunemap_overview_server}
\alias{immunemap_overview_server}
\title{overview page for immune map application}
\usage{
immunemap_overview_server(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
overview page for immune map application
}
