% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_immunemap.R
\name{immunemap_overview_ui}
\alias{immunemap_overview_ui}
\title{Create overview page for immune map application}
\usage{
immunemap_overview_ui(id, ...)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{...}{dots - additional arguments (if any) to be passed to sub-modules}
}
\description{
Create overview page for immune map application
}
