% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_ImmuneMapApp.R
\name{ImmuneMapAppManager}
\alias{ImmuneMapAppManager}
\title{R6 Class to manage the Immune Map application - subclass of the core TrisomExplorerAppManager  -- DEPRECATED --}
\description{
subclass of the core TrisomExplorerAppManager - inherits core attributes and methods
}
\section{Super class}{
\code{\link[TrisomExploreR:TrisomExplorerAppManager]{TrisomExploreR::TrisomExplorerAppManager}} -> \code{ImmuneMapAppManager}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{analysis}}{- string -}

\item{\code{analysisMetadata}}{- tibble -}

\item{\code{CellType}}{- tibble - analysis choices for UI input}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImmuneMapAppManager-new}{\code{ImmuneMapAppManager$new()}}
\item \href{#method-ImmuneMapAppManager-clone}{\code{ImmuneMapAppManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmuneMapAppManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImmuneMapAppManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of a ImmuneMapAppManager
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmuneMapAppManager$new(ApplicationId, remoteDB, localDB)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApplicationId}}{- string - application id}

\item{\code{remoteDB}}{R6 class - query manager for remote database queries}

\item{\code{localDB}}{R6 class - query manager for local database queries}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmuneMapAppManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImmuneMapAppManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmuneMapAppManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
