% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_ImmuneMapCorrelates.R
\name{ImmuneMapCorrelatesManager}
\alias{ImmuneMapCorrelatesManager}
\title{R6 Class to manage Correlates Analysis for Immune Maps data}
\description{
Enables analysis of cross-omics correlations for Immune Maps
Subclass of CorrelatesManager
}
\section{Super class}{
\code{\link[TrisomExploreR:CorrelatesManager]{TrisomExploreR::CorrelatesManager}} -> \code{ImmuneMapCorrelatesManager}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImmuneMapCorrelatesManager-new}{\code{ImmuneMapCorrelatesManager$new()}}
\item \href{#method-ImmuneMapCorrelatesManager-getQueryAnalytes}{\code{ImmuneMapCorrelatesManager$getQueryAnalytes()}}
\item \href{#method-ImmuneMapCorrelatesManager-getVolcanoPlot}{\code{ImmuneMapCorrelatesManager$getVolcanoPlot()}}
\item \href{#method-ImmuneMapCorrelatesManager-clone}{\code{ImmuneMapCorrelatesManager$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="addGSEAInputClass"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-addGSEAInputClass'><code>TrisomExploreR::CorrelatesManager$addGSEAInputClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="annotate_volcano_point"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-annotate_volcano_point'><code>TrisomExploreR::CorrelatesManager$annotate_volcano_point()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="getAnalyteData"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-getAnalyteData'><code>TrisomExploreR::CorrelatesManager$getAnalyteData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="getAnalytePlot"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-getAnalytePlot'><code>TrisomExploreR::CorrelatesManager$getAnalytePlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="getComparisonPlatforms"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-getComparisonPlatforms'><code>TrisomExploreR::CorrelatesManager$getComparisonPlatforms()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="getFormattedAnalyteSummaryData"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-getFormattedAnalyteSummaryData'><code>TrisomExploreR::CorrelatesManager$getFormattedAnalyteSummaryData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="getFormattedVolcanoSummaryData"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-getFormattedVolcanoSummaryData'><code>TrisomExploreR::CorrelatesManager$getFormattedVolcanoSummaryData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="getGSEAData"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-getGSEAData'><code>TrisomExploreR::CorrelatesManager$getGSEAData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="getGSEAEnrichmentPlot"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-getGSEAEnrichmentPlot'><code>TrisomExploreR::CorrelatesManager$getGSEAEnrichmentPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="getGSEAPathwayData"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-getGSEAPathwayData'><code>TrisomExploreR::CorrelatesManager$getGSEAPathwayData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="getGSEAPlot"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-getGSEAPlot'><code>TrisomExploreR::CorrelatesManager$getGSEAPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="getQueryPlatforms"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-getQueryPlatforms'><code>TrisomExploreR::CorrelatesManager$getQueryPlatforms()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="getVolcanoMultiSelectText"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-getVolcanoMultiSelectText'><code>TrisomExploreR::CorrelatesManager$getVolcanoMultiSelectText()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="getVolcanoSummaryData"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-getVolcanoSummaryData'><code>TrisomExploreR::CorrelatesManager$getVolcanoSummaryData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="get_analyte_plot_method"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-get_analyte_plot_method'><code>TrisomExploreR::CorrelatesManager$get_analyte_plot_method()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="updateAnalyteAttributes"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-updateAnalyteAttributes'><code>TrisomExploreR::CorrelatesManager$updateAnalyteAttributes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="CorrelatesManager" data-id="validate_volcano_plot"><a href='../../TrisomExploreR/html/CorrelatesManager.html#method-CorrelatesManager-validate_volcano_plot'><code>TrisomExploreR::CorrelatesManager$validate_volcano_plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmuneMapCorrelatesManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImmuneMapCorrelatesManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of ImmuneMapCorrelatesManager object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmuneMapCorrelatesManager$new(
  applicationName,
  id,
  namespace_config,
  remoteDB,
  localDB
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{applicationName}}{string - applicationName}

\item{\code{id}}{string - namespace for this class}

\item{\code{namespace_config}}{tibble - configuration values for this namespace instance of the object}

\item{\code{remoteDB}}{R6 class to manage remote database queries}

\item{\code{localDB}}{R6 class to manange local database queries}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ImmuneMapCorrelatesManager` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmuneMapCorrelatesManager-getQueryAnalytes"></a>}}
\if{latex}{\out{\hypertarget{method-ImmuneMapCorrelatesManager-getQueryAnalytes}{}}}
\subsection{Method \code{getQueryAnalytes()}}{
Get query analyte choices for the chosen QueryPlatform
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmuneMapCorrelatesManager$getQueryAnalytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmuneMapCorrelatesManager-getVolcanoPlot"></a>}}
\if{latex}{\out{\hypertarget{method-ImmuneMapCorrelatesManager-getVolcanoPlot}{}}}
\subsection{Method \code{getVolcanoPlot()}}{
Get volcano plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmuneMapCorrelatesManager$getVolcanoPlot(.data, ns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{tibble - data for volcano plot}

\item{\code{ns}}{- namespace to apply to plot object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmuneMapCorrelatesManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImmuneMapCorrelatesManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmuneMapCorrelatesManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
