% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_ImmuneMapFeatureAnalysis.R
\name{ImmuneMapFeatureAnalysisManager}
\alias{ImmuneMapFeatureAnalysisManager}
\title{R6 Class to  manage Immune Map specific Feature Analysis}
\description{
subclass of FeatureAnalysisManager - overrides several class methods
}
\section{Super class}{
\code{\link[TrisomExploreR:FeatureAnalysisManager]{TrisomExploreR::FeatureAnalysisManager}} -> \code{ImmuneMapFeatureAnalysisManager}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{analysisMetadata}}{- tibble - defines UI analysis choice metadata}

\item{\code{analysisChoices}}{- tibble - values available for analysis choices on UI side}

\item{\code{Analysis}}{- string - chosen analysis}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ImmuneMapFeatureAnalysisManager-new}{\code{ImmuneMapFeatureAnalysisManager$new()}}
\item \href{#method-ImmuneMapFeatureAnalysisManager-getAnalysisChoices}{\code{ImmuneMapFeatureAnalysisManager$getAnalysisChoices()}}
\item \href{#method-ImmuneMapFeatureAnalysisManager-getCellTypes}{\code{ImmuneMapFeatureAnalysisManager$getCellTypes()}}
\item \href{#method-ImmuneMapFeatureAnalysisManager-getAnalysisInputChoices}{\code{ImmuneMapFeatureAnalysisManager$getAnalysisInputChoices()}}
\item \href{#method-ImmuneMapFeatureAnalysisManager-getKaryotypeChoices}{\code{ImmuneMapFeatureAnalysisManager$getKaryotypeChoices()}}
\item \href{#method-ImmuneMapFeatureAnalysisManager-getBaseData}{\code{ImmuneMapFeatureAnalysisManager$getBaseData()}}
\item \href{#method-ImmuneMapFeatureAnalysisManager-getVolcanoSummaryData}{\code{ImmuneMapFeatureAnalysisManager$getVolcanoSummaryData()}}
\item \href{#method-ImmuneMapFeatureAnalysisManager-clone}{\code{ImmuneMapFeatureAnalysisManager$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="addGSEAInputClass"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-addGSEAInputClass'><code>TrisomExploreR::FeatureAnalysisManager$addGSEAInputClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="addInputSpecialClass"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-addInputSpecialClass'><code>TrisomExploreR::FeatureAnalysisManager$addInputSpecialClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getAnalyteData"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getAnalyteData'><code>TrisomExploreR::FeatureAnalysisManager$getAnalyteData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getAnalytePlot"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getAnalytePlot'><code>TrisomExploreR::FeatureAnalysisManager$getAnalytePlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getConditionTree"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getConditionTree'><code>TrisomExploreR::FeatureAnalysisManager$getConditionTree()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getCovariateChoices"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getCovariateChoices'><code>TrisomExploreR::FeatureAnalysisManager$getCovariateChoices()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getDisabledInputClass"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getDisabledInputClass'><code>TrisomExploreR::FeatureAnalysisManager$getDisabledInputClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getFormattedAnalyteSummaryData"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getFormattedAnalyteSummaryData'><code>TrisomExploreR::FeatureAnalysisManager$getFormattedAnalyteSummaryData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getFormattedVolcanoSummaryData"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getFormattedVolcanoSummaryData'><code>TrisomExploreR::FeatureAnalysisManager$getFormattedVolcanoSummaryData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getGSEAData"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getGSEAData'><code>TrisomExploreR::FeatureAnalysisManager$getGSEAData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getGSEAEnrichmentPlot"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getGSEAEnrichmentPlot'><code>TrisomExploreR::FeatureAnalysisManager$getGSEAEnrichmentPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getGSEAPathwayData"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getGSEAPathwayData'><code>TrisomExploreR::FeatureAnalysisManager$getGSEAPathwayData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getGSEAPlot"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getGSEAPlot'><code>TrisomExploreR::FeatureAnalysisManager$getGSEAPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getGetDataButtonClass"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getGetDataButtonClass'><code>TrisomExploreR::FeatureAnalysisManager$getGetDataButtonClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getHiddenInputClass"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getHiddenInputClass'><code>TrisomExploreR::FeatureAnalysisManager$getHiddenInputClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getSelectedConditionList"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getSelectedConditionList'><code>TrisomExploreR::FeatureAnalysisManager$getSelectedConditionList()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getStudies"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getStudies'><code>TrisomExploreR::FeatureAnalysisManager$getStudies()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getVolcanoPlot"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getVolcanoPlot'><code>TrisomExploreR::FeatureAnalysisManager$getVolcanoPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="get_analyte_plot_method"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-get_analyte_plot_method'><code>TrisomExploreR::FeatureAnalysisManager$get_analyte_plot_method()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="get_analyte_plot_title"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-get_analyte_plot_title'><code>TrisomExploreR::FeatureAnalysisManager$get_analyte_plot_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="get_analyte_plot_x_axis_label"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-get_analyte_plot_x_axis_label'><code>TrisomExploreR::FeatureAnalysisManager$get_analyte_plot_x_axis_label()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="updateAnalyteAttributes"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-updateAnalyteAttributes'><code>TrisomExploreR::FeatureAnalysisManager$updateAnalyteAttributes()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmuneMapFeatureAnalysisManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-ImmuneMapFeatureAnalysisManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of a ImmuneMapFeatureAnalysisManager
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmuneMapFeatureAnalysisManager$new(
  applicationName,
  id,
  namespace_config,
  remoteDB,
  localDB
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{applicationName}}{string - name of application}

\item{\code{id}}{string - namespace for this instance}

\item{\code{namespace_config}}{list - configurations for this namespace}

\item{\code{remoteDB}}{R6 class - query manager for remote database queries}

\item{\code{localDB}}{R6 class - query manager for local database queries}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmuneMapFeatureAnalysisManager-getAnalysisChoices"></a>}}
\if{latex}{\out{\hypertarget{method-ImmuneMapFeatureAnalysisManager-getAnalysisChoices}{}}}
\subsection{Method \code{getAnalysisChoices()}}{
#' helper function to get analysis choices for UI inputs based on chosen study
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmuneMapFeatureAnalysisManager$getAnalysisChoices(study)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study}}{- string - chosen study}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmuneMapFeatureAnalysisManager-getCellTypes"></a>}}
\if{latex}{\out{\hypertarget{method-ImmuneMapFeatureAnalysisManager-getCellTypes}{}}}
\subsection{Method \code{getCellTypes()}}{
helper function to get cell type choices for UI inputs based on chosen study
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmuneMapFeatureAnalysisManager$getCellTypes(study)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{study}}{- string - chosen study}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmuneMapFeatureAnalysisManager-getAnalysisInputChoices"></a>}}
\if{latex}{\out{\hypertarget{method-ImmuneMapFeatureAnalysisManager-getAnalysisInputChoices}{}}}
\subsection{Method \code{getAnalysisInputChoices()}}{
helper function to get analysis input choices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmuneMapFeatureAnalysisManager$getAnalysisInputChoices()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmuneMapFeatureAnalysisManager-getKaryotypeChoices"></a>}}
\if{latex}{\out{\hypertarget{method-ImmuneMapFeatureAnalysisManager-getKaryotypeChoices}{}}}
\subsection{Method \code{getKaryotypeChoices()}}{
helper function to get Karyotype input options based on namespace
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmuneMapFeatureAnalysisManager$getKaryotypeChoices(karyotypes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{karyotypes}}{string vector of karyotypes for input widget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmuneMapFeatureAnalysisManager-getBaseData"></a>}}
\if{latex}{\out{\hypertarget{method-ImmuneMapFeatureAnalysisManager-getBaseData}{}}}
\subsection{Method \code{getBaseData()}}{
Get / set sample level data with filers applied
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmuneMapFeatureAnalysisManager$getBaseData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmuneMapFeatureAnalysisManager-getVolcanoSummaryData"></a>}}
\if{latex}{\out{\hypertarget{method-ImmuneMapFeatureAnalysisManager-getVolcanoSummaryData}{}}}
\subsection{Method \code{getVolcanoSummaryData()}}{
Set Volcano Summary Data along with other volcano plot properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmuneMapFeatureAnalysisManager$getVolcanoSummaryData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ImmuneMapFeatureAnalysisManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ImmuneMapFeatureAnalysisManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ImmuneMapFeatureAnalysisManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
