% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_utils.R
\name{load_modules}
\alias{load_modules}
\title{Utility server function to dynamically load classes and modules}
\usage{
load_modules(id, module_config)
}
\arguments{
\item{id}{namespace for this module instance}

\item{module_config}{list - list of configurations for module}
}
\description{
Utility server function to dynamically load classes and modules
}
