% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_metabolome.R
\name{metabolome_overview_server}
\alias{metabolome_overview_server}
\title{Create overview page for metabolome application}
\usage{
metabolome_overview_server(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create overview page for metabolome application
}
