% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_metabolome.R
\name{metabolome_overview_ui}
\alias{metabolome_overview_ui}
\title{Create overview page for metabolome application}
\usage{
metabolome_overview_ui(id, ...)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{...}{dots - additional arguments (if any) to be passed to sub-modules}
}
\description{
Create overview page for metabolome application
}
