% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_DatabaseManagers.R
\name{ODBCQueryManager}
\alias{ODBCQueryManager}
\title{R6 Class to manage ODBC Database queries - subclass of OBBCQueryManager}
\description{
Manage ODBC Database queries
}
\section{Super class}{
\code{\link[TrisomExploreR:ODBCConnectionManager]{TrisomExploreR::ODBCConnectionManager}} -> \code{ODBCQueryManager}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{queryString}}{- string - Parameterized SQL Query to execute against target database}

\item{\code{parameters}}{- tibble - tibble of parameter names and values}

\item{\code{convertFactorsToStrings}}{- logical - whether to convert all factors to strings}

\item{\code{data}}{- tibble - query result formatted as tibble}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ODBCQueryManager-new}{\code{ODBCQueryManager$new()}}
\item \href{#method-ODBCQueryManager-getQuery}{\code{ODBCQueryManager$getQuery()}}
\item \href{#method-ODBCQueryManager-clone}{\code{ODBCQueryManager$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="ODBCConnectionManager" data-id="connect"><a href='../../TrisomExploreR/html/ODBCConnectionManager.html#method-ODBCConnectionManager-connect'><code>TrisomExploreR::ODBCConnectionManager$connect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="ODBCConnectionManager" data-id="disconnect"><a href='../../TrisomExploreR/html/ODBCConnectionManager.html#method-ODBCConnectionManager-disconnect'><code>TrisomExploreR::ODBCConnectionManager$disconnect()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ODBCQueryManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-ODBCQueryManager-new}{}}}
\subsection{Method \code{new()}}{
helper function to set parameters tibble


helper function to properly format query result


Create a new instance of ODBCQueryManager object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ODBCQueryManager$new(conn_args)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{conn_args}}{list - list of connection arguments to connect to database}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `ODBCQueryManager` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ODBCQueryManager-getQuery"></a>}}
\if{latex}{\out{\hypertarget{method-ODBCQueryManager-getQuery}{}}}
\subsection{Method \code{getQuery()}}{
Execute parameterized query against target database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ODBCQueryManager$getQuery(
  queryString,
  parameters,
  convertFactorsToStrings = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{queryString}}{- string - parameterized sql query string}

\item{\code{parameters}}{- tibble - parameter names and values}

\item{\code{convertFactorsToStrings}}{- logical - whether to convert factors to strings}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ODBCQueryManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ODBCQueryManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ODBCQueryManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
