% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_precalc_feature_analysis.R
\name{precalc_feature_analysis_server}
\alias{precalc_feature_analysis_server}
\title{Server-side logic / processing for TrisomExploreR precalculated feature analysis - DESeq2}
\usage{
precalc_feature_analysis_server(id, r6, ...)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic to be utilized by all sub-modules}

\item{...}{dots - additional arguments (if any) to be passed to sub-modules}
}
\description{
Server-side logic / processing for TrisomExploreR precalculated feature analysis - DESeq2
}
