% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_precalc_feature_analysis.R
\name{precalc_feature_analysis_ui}
\alias{precalc_feature_analysis_ui}
\title{Create standard layout / skeleton / sub-modules for TrisomExploreR precalculated feature analysis - DESeq2 
Follows same skeleton as standard feature analysis but with precalc specific input module}
\usage{
precalc_feature_analysis_ui(id, ...)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{...}{dots - additional arguments (if any) to be passed to sub-modules}
}
\description{
Create standard layout / skeleton / sub-modules for TrisomExploreR precalculated feature analysis - DESeq2 
Follows same skeleton as standard feature analysis but with precalc specific input module
}
