% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_PreCalcFeatureAnalysis.R
\name{PreCalcFeatureAnalysisManager}
\alias{PreCalcFeatureAnalysisManager}
\title{R6 Class to  manage precalculated Feature Analysis - DESeq2}
\description{
subclass of FeatureAnalysisManager - overrides several class methods
}
\section{Super class}{
\code{\link[TrisomExploreR:FeatureAnalysisManager]{TrisomExploreR::FeatureAnalysisManager}} -> \code{PreCalcFeatureAnalysisManager}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Age}}{- numeric vector of chosen ages}

\item{\code{Sex}}{- character vecotr of chosen sex(s)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PreCalcFeatureAnalysisManager-new}{\code{PreCalcFeatureAnalysisManager$new()}}
\item \href{#method-PreCalcFeatureAnalysisManager-getKaryotypeChoices}{\code{PreCalcFeatureAnalysisManager$getKaryotypeChoices()}}
\item \href{#method-PreCalcFeatureAnalysisManager-getVolcanoSummaryData}{\code{PreCalcFeatureAnalysisManager$getVolcanoSummaryData()}}
\item \href{#method-PreCalcFeatureAnalysisManager-getAnalyteData}{\code{PreCalcFeatureAnalysisManager$getAnalyteData()}}
\item \href{#method-PreCalcFeatureAnalysisManager-getGSEAData}{\code{PreCalcFeatureAnalysisManager$getGSEAData()}}
\item \href{#method-PreCalcFeatureAnalysisManager-clone}{\code{PreCalcFeatureAnalysisManager$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="addGSEAInputClass"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-addGSEAInputClass'><code>TrisomExploreR::FeatureAnalysisManager$addGSEAInputClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="addInputSpecialClass"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-addInputSpecialClass'><code>TrisomExploreR::FeatureAnalysisManager$addInputSpecialClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="annotate_volcano_point"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-annotate_volcano_point'><code>TrisomExploreR::FeatureAnalysisManager$annotate_volcano_point()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getAnalytePlot"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getAnalytePlot'><code>TrisomExploreR::FeatureAnalysisManager$getAnalytePlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getBaseData"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getBaseData'><code>TrisomExploreR::FeatureAnalysisManager$getBaseData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getConditionTree"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getConditionTree'><code>TrisomExploreR::FeatureAnalysisManager$getConditionTree()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getCovariateChoices"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getCovariateChoices'><code>TrisomExploreR::FeatureAnalysisManager$getCovariateChoices()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getDisabledInputClass"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getDisabledInputClass'><code>TrisomExploreR::FeatureAnalysisManager$getDisabledInputClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getFormattedAnalyteSummaryData"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getFormattedAnalyteSummaryData'><code>TrisomExploreR::FeatureAnalysisManager$getFormattedAnalyteSummaryData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getFormattedVolcanoSummaryData"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getFormattedVolcanoSummaryData'><code>TrisomExploreR::FeatureAnalysisManager$getFormattedVolcanoSummaryData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getGSEAEnrichmentPlot"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getGSEAEnrichmentPlot'><code>TrisomExploreR::FeatureAnalysisManager$getGSEAEnrichmentPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getGSEAPathwayData"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getGSEAPathwayData'><code>TrisomExploreR::FeatureAnalysisManager$getGSEAPathwayData()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getGSEAPlot"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getGSEAPlot'><code>TrisomExploreR::FeatureAnalysisManager$getGSEAPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getGetDataButtonClass"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getGetDataButtonClass'><code>TrisomExploreR::FeatureAnalysisManager$getGetDataButtonClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getHiddenInputClass"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getHiddenInputClass'><code>TrisomExploreR::FeatureAnalysisManager$getHiddenInputClass()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getSelectedConditionList"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getSelectedConditionList'><code>TrisomExploreR::FeatureAnalysisManager$getSelectedConditionList()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getStudies"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getStudies'><code>TrisomExploreR::FeatureAnalysisManager$getStudies()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getVolcanoMultiSelectText"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getVolcanoMultiSelectText'><code>TrisomExploreR::FeatureAnalysisManager$getVolcanoMultiSelectText()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="getVolcanoPlot"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-getVolcanoPlot'><code>TrisomExploreR::FeatureAnalysisManager$getVolcanoPlot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="get_analyte_plot_method"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-get_analyte_plot_method'><code>TrisomExploreR::FeatureAnalysisManager$get_analyte_plot_method()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="get_analyte_plot_title"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-get_analyte_plot_title'><code>TrisomExploreR::FeatureAnalysisManager$get_analyte_plot_title()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="get_analyte_plot_x_axis_label"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-get_analyte_plot_x_axis_label'><code>TrisomExploreR::FeatureAnalysisManager$get_analyte_plot_x_axis_label()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="updateAnalyteAttributes"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-updateAnalyteAttributes'><code>TrisomExploreR::FeatureAnalysisManager$updateAnalyteAttributes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TrisomExploreR" data-topic="FeatureAnalysisManager" data-id="validate_volcano_plot"><a href='../../TrisomExploreR/html/FeatureAnalysisManager.html#method-FeatureAnalysisManager-validate_volcano_plot'><code>TrisomExploreR::FeatureAnalysisManager$validate_volcano_plot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PreCalcFeatureAnalysisManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-PreCalcFeatureAnalysisManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of a PreCalcFeatureAnalysisManager
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreCalcFeatureAnalysisManager$new(
  applicationName,
  id,
  namespace_config,
  remoteDB,
  localDB
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{applicationName}}{string - name of application}

\item{\code{id}}{string - namespace for this instance}

\item{\code{namespace_config}}{list - configurations for this namespace}

\item{\code{remoteDB}}{R6 class - query manager for remote database queries}

\item{\code{localDB}}{R6 class - query manager for local database queries}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PreCalcFeatureAnalysisManager-getKaryotypeChoices"></a>}}
\if{latex}{\out{\hypertarget{method-PreCalcFeatureAnalysisManager-getKaryotypeChoices}{}}}
\subsection{Method \code{getKaryotypeChoices()}}{
helper function to get Karyotype input options based on namespace
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreCalcFeatureAnalysisManager$getKaryotypeChoices(karyotypes)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{karyotypes}}{string vector of karyotypes for input widget}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PreCalcFeatureAnalysisManager-getVolcanoSummaryData"></a>}}
\if{latex}{\out{\hypertarget{method-PreCalcFeatureAnalysisManager-getVolcanoSummaryData}{}}}
\subsection{Method \code{getVolcanoSummaryData()}}{
Set Volcano Summary Data along with other volcano plot properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreCalcFeatureAnalysisManager$getVolcanoSummaryData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PreCalcFeatureAnalysisManager-getAnalyteData"></a>}}
\if{latex}{\out{\hypertarget{method-PreCalcFeatureAnalysisManager-getAnalyteData}{}}}
\subsection{Method \code{getAnalyteData()}}{
get sample level data for selected analyte(s)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreCalcFeatureAnalysisManager$getAnalyteData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PreCalcFeatureAnalysisManager-getGSEAData"></a>}}
\if{latex}{\out{\hypertarget{method-PreCalcFeatureAnalysisManager-getGSEAData}{}}}
\subsection{Method \code{getGSEAData()}}{
Get GSEA data
### NO NEED TO PENALIZE SINCE THE MODEL IS DESEQ**
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreCalcFeatureAnalysisManager$getGSEAData()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PreCalcFeatureAnalysisManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PreCalcFeatureAnalysisManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PreCalcFeatureAnalysisManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
