% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_proteome.R
\name{proteome_overview_server}
\alias{proteome_overview_server}
\title{overview page for proteome application}
\usage{
proteome_overview_server(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
overview page for proteome application
}
