% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_proteome.R
\name{proteome_overview_ui}
\alias{proteome_overview_ui}
\title{Create overview page for proteome application}
\usage{
proteome_overview_ui(id, ...)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{...}{dots - additional arguments (if any) to be passed to sub-modules}
}
\description{
Create overview page for proteome application
}
