% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{purge_plot}
\alias{purge_plot}
\title{Helper function to purge plotly plot
Calls custom JS scripts which can be found in `inst/assets/js/script.js`}
\usage{
purge_plot(session, ns, plot_name, r6)
}
\arguments{
\item{session}{- shiny session object}

\item{ns}{- NS - namespace - not currently used}

\item{plot_name}{- string - name of target plot}

\item{r6}{- R6 class - R6 class with namespace logic}
}
\description{
Helper function to purge plotly plot
Calls custom JS scripts which can be found in `inst/assets/js/script.js`
}
