% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_DatabaseManagers.R
\name{SQLiteConnectionManager}
\alias{SQLiteConnectionManager}
\title{R6 Class to manage SQLite Database connections}
\description{
Manage SQLite Database connections
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{connection_open}}{- logical - whether the connection is currently is open or not}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SQLiteConnectionManager-new}{\code{SQLiteConnectionManager$new()}}
\item \href{#method-SQLiteConnectionManager-connect}{\code{SQLiteConnectionManager$connect()}}
\item \href{#method-SQLiteConnectionManager-disconnect}{\code{SQLiteConnectionManager$disconnect()}}
\item \href{#method-SQLiteConnectionManager-clone}{\code{SQLiteConnectionManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SQLiteConnectionManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-SQLiteConnectionManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of SQLiteConnectionManager object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLiteConnectionManager$new(filepath)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filepath}}{string - path to `.sqlite` database file}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SQLiteConnectionManager` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SQLiteConnectionManager-connect"></a>}}
\if{latex}{\out{\hypertarget{method-SQLiteConnectionManager-connect}{}}}
\subsection{Method \code{connect()}}{
Connect to target database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLiteConnectionManager$connect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SQLiteConnectionManager-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-SQLiteConnectionManager-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Disconnect from target database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLiteConnectionManager$disconnect()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SQLiteConnectionManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SQLiteConnectionManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SQLiteConnectionManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
