% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{states_shp}
\alias{states_shp}
\title{states_shp}
\format{
An sf (simple features) dataset containing spatial data for 52 states within the United states. The dataset contains the following attributes:
}
\usage{
data("states_shp")
}
\description{
This geospatial dataset defines the current Census State and Equivalent National geospatial entities. It was generated using data sourced from
US Census data, which is publically available here: \url{https://www.census.gov/geographies/mapping-files/2013/geo/carto-boundary-file.html}
}
\details{
\itemize{
  \item STATEFP, Current state Federal Information Processing Series (FIPS) code
  \item STATENS, Current state ANSI code
  \item AFFGEOID, American FactFinder summary level code + geovariant code + '00US' + GEOID
  \item GEOID, State identifier; state FIPS code
  \item STUSPS, Current United States Postal Service state abbreviation
  \item NAME, Current state name
  \item LSAD, Current legal/statistical area description code for state
  \item ALAND, Current land area (square meters)
  \item AWATER, Current water area (square meters)
  \item geometry, mutlipolygon - list of geographical elements defining shape
}
}
\keyword{datasets}
