% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{toggle_GSEA_volcano_plot_trace}
\alias{toggle_GSEA_volcano_plot_trace}
\title{Helper function to clone volcano plot traces for pathway / GSEA analysis
Calls custom JS scripts which can be found in `inst/assets/js/script.js`}
\usage{
toggle_GSEA_volcano_plot_trace(
  session,
  ns,
  plot_name = "VolcanoPlot",
  r6,
  action = "add"
)
}
\arguments{
\item{session}{- shiny session object - session for target plot}

\item{ns}{- NS - namespace - not currently used}

\item{plot_name}{- string - name of target plot - defaults to "VolcanoPlot"}

\item{r6}{- R6 class - R6 class with namespace logic}

\item{action}{- string - if "add", will clone and add traces to volcano plot, otherwise will remove traces}
}
\description{
Helper function to clone volcano plot traces for pathway / GSEA analysis
Calls custom JS scripts which can be found in `inst/assets/js/script.js`
}
