% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_transcriptome.R
\name{transcriptome_overview_server}
\alias{transcriptome_overview_server}
\title{overview page for transcriptome application}
\usage{
transcriptome_overview_server(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
overview page for transcriptome application
}
