% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_TranscriptomeApp.R
\name{TranscriptomeAppManager}
\alias{TranscriptomeAppManager}
\title{R6 Class to manage TrisomExploreR Transcriptome application  -- DEPRECATED --}
\description{
subclass of TrisomExplorerAppManager
}
\section{Super class}{
\code{\link[TrisomExploreR:TrisomExplorerAppManager]{TrisomExploreR::TrisomExplorerAppManager}} -> \code{TranscriptomeAppManager}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TranscriptomeAppManager-new}{\code{TranscriptomeAppManager$new()}}
\item \href{#method-TranscriptomeAppManager-clone}{\code{TranscriptomeAppManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranscriptomeAppManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-TranscriptomeAppManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of a TranscriptomeAppManager
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranscriptomeAppManager$new(ApplicationId, remoteDB, localDB)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApplicationId}}{- string - application id}

\item{\code{remoteDB}}{R6 class - query manager for remote database queries}

\item{\code{localDB}}{R6 class - query manager for local database queries}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranscriptomeAppManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TranscriptomeAppManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranscriptomeAppManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
