% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_TrisomExplorerApp.R
\name{TrisomExplorerAppManager}
\alias{TrisomExplorerAppManager}
\title{R6 Class to manage core TrisomExploreR applications}
\description{
R6 Class to manage core TrisomExploreR applications
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{app_config}}{- list -}

\item{\code{module_config}}{- list -}

\item{\code{analysis_config}}{- list -}

\item{\code{input_config}}{- list -}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TrisomExplorerAppManager-new}{\code{TrisomExplorerAppManager$new()}}
\item \href{#method-TrisomExplorerAppManager-clone}{\code{TrisomExplorerAppManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrisomExplorerAppManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-TrisomExplorerAppManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of a TrisomExplorerAppManager
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrisomExplorerAppManager$new(ApplicationId, remoteDB, localDB)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ApplicationId}}{- string - application id}

\item{\code{remoteDB}}{R6 class - query manager for remote database queries}

\item{\code{localDB}}{R6 class - query manager for local database queries}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TrisomExplorerAppManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TrisomExplorerAppManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TrisomExplorerAppManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
