% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_volcano_datatable.R
\name{volcano_data_table_server}
\alias{volcano_data_table_server}
\title{Server logic for volcano summary data table output for TrisomExploreR feature analysis#'}
\usage{
volcano_data_table_server(id, r6)
}
\arguments{
\item{id}{- string - id for this module namespace}

\item{r6}{- R6 class defining server-side logic}
}
\description{
Server logic for volcano summary data table output for TrisomExploreR feature analysis#'
}
