% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_volcano_datatable.R
\name{volcano_data_table_ui}
\alias{volcano_data_table_ui}
\title{Create volcano summary data table output for TrisomExploreR feature analysis}
\usage{
volcano_data_table_ui(id)
}
\arguments{
\item{id}{- string - id for this module namespace}
}
\description{
Create volcano summary data table output for TrisomExploreR feature analysis
}
