% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_volcano_plot_analyte.R
\name{volcano_plot_analyte_input_server}
\alias{volcano_plot_analyte_input_server}
\title{Server side processing / logic for analyte input for volcano plot}
\usage{
volcano_plot_analyte_input_server(id, r6, parent)
}
\arguments{
\item{id}{namespace for this module instance}

\item{r6}{r6 class for data management}

\item{parent}{shiny session - parent session}
}
\description{
Server side processing / logic for analyte input for volcano plot
}
