% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs_volcano_plot_analyte.R
\name{volcano_plot_analyte_input_ui}
\alias{volcano_plot_analyte_input_ui}
\title{Create analyte inputs for volcano plot}
\usage{
volcano_plot_analyte_input_ui(id)
}
\arguments{
\item{id}{namespace for this module instance}
}
\description{
Create analyte inputs for volcano plot
}
