% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_volcano.R
\name{volcano_plot_server}
\alias{volcano_plot_server}
\title{server side logic / processing for volcano plot ui module}
\usage{
volcano_plot_server(id, r6, ...)
}
\arguments{
\item{id}{namespace for this module instance}

\item{r6}{r6 class for data management}

\item{...}{dots - additional arguments passed to other submodules}
}
\description{
server side logic / processing for volcano plot ui module
}
