% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_Correlates.R
\name{CorrelatesManager}
\alias{CorrelatesManager}
\title{R6 Class to manage Correlates Analysis}
\description{
Enables analysis of cross-omics correlations
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{applicationName}}{- string - name of application}

\item{\code{ApplicationId}}{- string - application ID}

\item{\code{namespace}}{- string - namespace for this instance}

\item{\code{remoteDB}}{- R6 class to manage remote database queries}

\item{\code{localDB}}{- R6 class to manage local database queries}

\item{\code{analysisType}}{- string - type of analysis for this instance/namespace}

\item{\code{FoldChangeVar}}{- string - name of variable indicating fold change or difference (log2FoldChange)}

\item{\code{SignificanceVariable}}{- string - name of variable indiciating significance value (p-value)}

\item{\code{Study}}{- string - selected study}

\item{\code{QueryPlatform}}{- string - selected query platform}

\item{\code{ComparisonPlatform}}{- string - selected comparison platform}

\item{\code{Experiment}}{- string - chosen experimentID}

\item{\code{QueryAnalyte}}{- string - chosen Query Analyte}

\item{\code{Analyte}}{- string - chosen Comparison Analyte}

\item{\code{Sex}}{- string vector - Sex values chosen for analysis}

\item{\code{Age}}{- numeric vector - Age values chosen fo analysis}

\item{\code{StatTest}}{- string - name of statistical test to apply for analysis (Linear Model, etc.)}

\item{\code{Covariates}}{- string vector - names of features to include as covariates in Linear Model analysis}

\item{\code{AdjustmentMethod}}{- string - name of multiple hypothesis correction method to apply to statistical output}

\item{\code{SignificanceLabel}}{- string - if adjusted, `q-value`, otherwise `p-value`}

\item{\code{AnalyteData}}{- tibble - sample level data for chosen analyte(s)}

\item{\code{AnalyteDataDownload}}{- deprecated?}

\item{\code{VolcanoSummaryData}}{- tibble - Fold Change summary data used for volcano plot}

\item{\code{volcanoSourceData}}{- tibble of formatted source data used for volcano plot - includes trace groups}

\item{\code{volcanoEventData}}{- tibble of click and selection data from volcano plot}

\item{\code{volcanoMultiSelectText}}{- string - text shown below volcano plot when multiple analytes are chosen}

\item{\code{HeatmapData}}{- tibble - data to use for heatmap plot when multiple analytes are chosen}

\item{\code{HeatmapText}}{- string vector - text used for hover on heatmap plot}

\item{\code{HeatmapLimit}}{- numeric - maxiumum abs correlation value used to determine upper/lower bounds of heatmap}

\item{\code{GSEAData}}{- list of ranks, hallmarks, and gsea results}

\item{\code{GSEAAnalytes}}{- character vector - matching analytes for for chosen GSEA pathway}

\item{\code{GSEATraceName}}{- string - name of chosen GSEA pathway}

\item{\code{GSEAGenesetName}}{- string - formatted version of GSEA Trace Name}

\item{\code{GSEAPathwayData}}{- pathway specific data for chosen GSEA pathway}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{AnalyteSearchName}}{- string - name of analyte used for external links}

\item{\code{CorrelationMeasureName}}{- string - name of correlation measure used in analysis - `rho` or `Fold Change`}

\item{\code{Adjusted}}{- logical - whether the statistical test includes multiple hypothesis correction or not}

\item{\code{ComparisonAnalytePlotStatAnnotation}}{- string - annotation label to show at top of volcano plot}

\item{\code{AnalytePlotMethod}}{- string - type of plot to show for chosen analyte}

\item{\code{CorrelationAnalytePlotTitle}}{- string - title to show for analyte plot}

\item{\code{QueryMeasurement}}{- string - measurement used for query analyte sample level data}

\item{\code{ComparisonMeasurement}}{- string - measurement used for comparison analyte sample level data}

\item{\code{QueryAnalyteLabel}}{- string - label used for query analyte}

\item{\code{ComparisonAnalyteLabel}}{- string - labe used for comparison analyte}

\item{\code{VolcanoSummaryDataXAxisLabel}}{- string - volcano plot x-axis}

\item{\code{VolcanoSummaryDataYAxisLabel}}{- string - volcano plot y-axis}

\item{\code{VolcanoSummaryMaxFoldChange}}{- numeric - maxiumum abs. value of fold change}

\item{\code{VolcanoPlotTitle}}{- string - title to show above volcano plot}

\item{\code{volcanoPlotExpectedTraceCount}}{- numeric - number of base traces present in the active volcano plot (usually between 1 - 3)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CorrelatesManager-new}{\code{CorrelatesManager$new()}}
\item \href{#method-CorrelatesManager-getQueryExperiments}{\code{CorrelatesManager$getQueryExperiments()}}
\item \href{#method-CorrelatesManager-addGSEAInputClass}{\code{CorrelatesManager$addGSEAInputClass()}}
\item \href{#method-CorrelatesManager-getQueryAnalytes}{\code{CorrelatesManager$getQueryAnalytes()}}
\item \href{#method-CorrelatesManager-getComparisonExperiments}{\code{CorrelatesManager$getComparisonExperiments()}}
\item \href{#method-CorrelatesManager-validate_volcano_plot}{\code{CorrelatesManager$validate_volcano_plot()}}
\item \href{#method-CorrelatesManager-set_correlation_source_data}{\code{CorrelatesManager$set_correlation_source_data()}}
\item \href{#method-CorrelatesManager-getVolcanoSummaryData}{\code{CorrelatesManager$getVolcanoSummaryData()}}
\item \href{#method-CorrelatesManager-getVolcanoPlot}{\code{CorrelatesManager$getVolcanoPlot()}}
\item \href{#method-CorrelatesManager-getFormattedVolcanoSummaryData}{\code{CorrelatesManager$getFormattedVolcanoSummaryData()}}
\item \href{#method-CorrelatesManager-updateAnalyteAttributes}{\code{CorrelatesManager$updateAnalyteAttributes()}}
\item \href{#method-CorrelatesManager-annotate_volcano_point}{\code{CorrelatesManager$annotate_volcano_point()}}
\item \href{#method-CorrelatesManager-getVolcanoMultiSelectText}{\code{CorrelatesManager$getVolcanoMultiSelectText()}}
\item \href{#method-CorrelatesManager-getAnalyteData}{\code{CorrelatesManager$getAnalyteData()}}
\item \href{#method-CorrelatesManager-getAnalytePlot}{\code{CorrelatesManager$getAnalytePlot()}}
\item \href{#method-CorrelatesManager-getFormattedAnalyteSummaryData}{\code{CorrelatesManager$getFormattedAnalyteSummaryData()}}
\item \href{#method-CorrelatesManager-getGSEAData}{\code{CorrelatesManager$getGSEAData()}}
\item \href{#method-CorrelatesManager-getGSEAPlot}{\code{CorrelatesManager$getGSEAPlot()}}
\item \href{#method-CorrelatesManager-getGSEAPathwayData}{\code{CorrelatesManager$getGSEAPathwayData()}}
\item \href{#method-CorrelatesManager-getGSEAEnrichmentPlot}{\code{CorrelatesManager$getGSEAEnrichmentPlot()}}
\item \href{#method-CorrelatesManager-clone}{\code{CorrelatesManager$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-new"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-new}{}}}
\subsection{Method \code{new()}}{
Create a new instance of CorrelatesManager object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$new(applicationName, id, namespace_config, remoteDB, localDB)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{applicationName}}{string - applicationName}

\item{\code{id}}{string - namespace for this class}

\item{\code{namespace_config}}{tibble - configuration values for this namespace instance of the object}

\item{\code{remoteDB}}{R6 class to manage remote database queries}

\item{\code{localDB}}{R6 class to manange local database queries}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `CorrelatesManager` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-getQueryExperiments"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-getQueryExperiments}{}}}
\subsection{Method \code{getQueryExperiments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$getQueryExperiments()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-addGSEAInputClass"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-addGSEAInputClass}{}}}
\subsection{Method \code{addGSEAInputClass()}}{
return hidden /shown / disabled / enabled class for GSEA button based on
 chosen comparison platform and whether or not Volcano Plot is rendered
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$addGSEAInputClass()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-getQueryAnalytes"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-getQueryAnalytes}{}}}
\subsection{Method \code{getQueryAnalytes()}}{
Get query analyte choices for the chosen QueryPlatform
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$getQueryAnalytes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
string vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-getComparisonExperiments"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-getComparisonExperiments}{}}}
\subsection{Method \code{getComparisonExperiments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$getComparisonExperiments()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-validate_volcano_plot"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-validate_volcano_plot}{}}}
\subsection{Method \code{validate_volcano_plot()}}{
validate that all inputs have been chosen to fetch volcano summary data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$validate_volcano_plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-set_correlation_source_data"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-set_correlation_source_data}{}}}
\subsection{Method \code{set_correlation_source_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$set_correlation_source_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-getVolcanoSummaryData"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-getVolcanoSummaryData}{}}}
\subsection{Method \code{getVolcanoSummaryData()}}{
Set Volcano Summary Data along with other volcano plot properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$getVolcanoSummaryData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-getVolcanoPlot"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-getVolcanoPlot}{}}}
\subsection{Method \code{getVolcanoPlot()}}{
Get Volcano Plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$getVolcanoPlot(.data, ns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- tibble - data for volcano plot}

\item{\code{ns}}{- namespace to apply to plot object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-getFormattedVolcanoSummaryData"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-getFormattedVolcanoSummaryData}{}}}
\subsection{Method \code{getFormattedVolcanoSummaryData()}}{
Get formatted volcano plot data for data table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$getFormattedVolcanoSummaryData(.data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{tibble - data for volcano plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-updateAnalyteAttributes"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-updateAnalyteAttributes}{}}}
\subsection{Method \code{updateAnalyteAttributes()}}{
update analyte fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$updateAnalyteAttributes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-annotate_volcano_point"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-annotate_volcano_point}{}}}
\subsection{Method \code{annotate_volcano_point()}}{
helper function to add annotation to volcano plot based on chosen analyte
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$annotate_volcano_point(plot_name, ns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot_name}}{string - name of target volcano plot}

\item{\code{ns}}{namespace to properly derive fully-qualified plot name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-getVolcanoMultiSelectText"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-getVolcanoMultiSelectText}{}}}
\subsection{Method \code{getVolcanoMultiSelectText()}}{
small helper function to get correct value for volcanoMultiSelectText
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$getVolcanoMultiSelectText()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-getAnalyteData"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-getAnalyteData}{}}}
\subsection{Method \code{getAnalyteData()}}{
set analyte sample level data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$getAnalyteData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-getAnalytePlot"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-getAnalytePlot}{}}}
\subsection{Method \code{getAnalytePlot()}}{
Get analyte plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$getAnalytePlot(.data, ns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{tibble - data for analyte plot}

\item{\code{ns}}{- namespace to apply to plot object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-getFormattedAnalyteSummaryData"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-getFormattedAnalyteSummaryData}{}}}
\subsection{Method \code{getFormattedAnalyteSummaryData()}}{
Get formatted analyte plot data for data table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$getFormattedAnalyteSummaryData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-getGSEAData"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-getGSEAData}{}}}
\subsection{Method \code{getGSEAData()}}{
Get GSEA data
** INCLUDES penalized calculation for ranks = `-log10pvalue` * CorrelationValue
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$getGSEAData()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of 3 tibbles:
* ranks - ranked analytes
* GSEA_hallmarks - geneset used
* gsea - gene set enrichment analysis
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-getGSEAPlot"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-getGSEAPlot}{}}}
\subsection{Method \code{getGSEAPlot()}}{
Get GSEA plot - top 25
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$getGSEAPlot(.data, ns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- data for plot}

\item{\code{ns}}{- namespace to apply to plot object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-getGSEAPathwayData"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-getGSEAPathwayData}{}}}
\subsection{Method \code{getGSEAPathwayData()}}{
Set selected GSEA pathway data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$getGSEAPathwayData(path_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path_name}}{- string - selected pathway name}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
none
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-getGSEAEnrichmentPlot"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-getGSEAEnrichmentPlot}{}}}
\subsection{Method \code{getGSEAEnrichmentPlot()}}{
Get selected GSEA pathway enrichment plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$getGSEAEnrichmentPlot(.data, ns)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data}}{- data used for plot}

\item{\code{ns}}{- namespace to apply to plot object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
plotly object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CorrelatesManager-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CorrelatesManager-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CorrelatesManager$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
