% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_Correlates.R,
%   R/manager_FeatureAnalysis.R
\name{getAnalytePlotMethod}
\alias{getAnalytePlotMethod}
\title{small helper function to determin plot method for instance/namespace}
\usage{
getAnalytePlotMethod(analysis_type, analyte_count)

getAnalytePlotMethod(analysis_type, analyte_count)
}
\arguments{
\item{analysis_type}{- string - analysis type for this instance/namespace}

\item{analyte_count}{int - number of selected analytes}

\item{analysisType}{- string - analysis type for this instance/namespace}

\item{analyteCount}{int - number of selected analytes}
}
\value{
string

string
}
\description{
small helper function to determin plot method for instance/namespace

small helper function to determine plot method for instance/namespace
}
