% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelatesManager.R
\name{getAnalytePlotMethod}
\alias{getAnalytePlotMethod}
\title{small helper function to determin plot method for instance/namespace}
\usage{
getAnalytePlotMethod(analysisType, analyteCount)
}
\arguments{
\item{analysisType}{- string - analysis type for this instance/namespace}

\item{analyteCount}{int - number of selected analytes}
}
\value{
string
}
\description{
small helper function to determin plot method for instance/namespace
}
