% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager_FeatureAnalysis.R
\name{getAnalytePlotTitle}
\alias{getAnalytePlotTitle}
\title{small helper function to generate plot title based on criteria}
\usage{
getAnalytePlotTitle(analysis_variable, plot_method, analyte, group_variable)
}
\arguments{
\item{analysis_variable}{- string - name of analysis variable}

\item{plot_method}{- string - one of boxplot, scatterplot, or heatmap}

\item{analyte}{- string vector - vector of chosen analytes}

\item{group_variable}{- string - grouping variable used in analysis (Karyotype)}
}
\value{
string
}
\description{
small helper function to generate plot title based on criteria
}
